% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl.R
\name{atc_classes}
\alias{atc_classes}
\title{Retrieve all ATC classes}
\usage{
atc_classes(verbose = getOption("verbose"))
}
\arguments{
\item{verbose}{logical; should a verbose output be printed on the console?
\donttest{
# Might fail if API is not available

atc <- atc_classes()
}}
}
\description{
Retrieve all available classes within the Anatomical Therapeutic Chemical
(ATC) classification system.
}
