% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotransformer.R
\name{biotransformer}
\alias{biotransformer}
\title{BioTransformer}
\usage{
biotransformer(jar = NULL, annotate = FALSE, btType = NULL,
  ismiles = NULL, task = NULL, formulas = NULL, masses = NULL,
  nsteps = 1, mTolerence = 0.01)
}
\arguments{
\item{jar}{Path to the biotransformer jar file}

\item{annotate}{(T/F) Search PuChem for each product, and store with CID and
synonyms, when available.}

\item{btType}{The type of description: Type of biotransformer
human super transformer (`superbio`, or `allHuman`), Environmental
microbial(`env`).}

\item{ismiles}{The input, which can be a SMILES string}

\item{task}{The task to be permed: pred for prediction, or cid for compound
identification}

\item{formulas}{Semicolon-separated list of formulas of compounds to identify}

\item{masses}{Semicolon-separated list of masses of compounds to identify}

\item{nsteps}{The number of steps for the prediction. This option can be set
by the user for the EC-based, CYP450, Phase II, and Environmental microbial
biotransformers. The default value is 1.}

\item{mTolerence}{Mass tolerance for metabolite identification (default is
0.01).}
}
\value{
A tibble containing predicted metabolite data
}
\description{
BioTransformer is a software
tool that predicts small molecule metabolism in mammals, their gut
microbiota, as well as the soil/aquatic microbiota. BioTransformer also
assists scientists in metabolite identification, based on the metabolism
prediction.
}
\details{
Download the jar file from: \url{https://bitbucket.org/djoumbou/biotransformerjar/src/master/}
Cite: Djoumbou-Feunang Y, Fiamoncini J, de la Fuente AG, Manach C, Greiner R,
and Wishart DS; BioTransformer: A Comprehensive Computational Tool for Small
Molecule Metabolism Prediction and Metabolite Identification; Journal of
Cheminformatics 201911:2; DOI: 10.1186/s13321-018-0324-5.
}
