% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alanwood.R
\name{build_aw_idx}
\alias{build_aw_idx}
\title{Function to build index}
\source{
\url{http://www.alanwood.net/pesticides}
}
\usage{
build_aw_idx(verbose = TRUE, force_build = FALSE)
}
\arguments{
\item{verbose}{logical; print message during processing to console?}

\item{force_build}{logical; force building a new index?}
}
\value{
a data.frame with three columns: cas, url, and name
}
\description{
This function builds an index of Alan Woods Compendium of Pesticides
\url{http://www.alanwood.net/pesticides} and saves it to \code{tempdir()}.
This is a utility function for \code{aw_query()} and will not be exported in
future releases.
}
\seealso{
\code{\link{aw_query}}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}
