% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{check_key}
\alias{check_key}
\alias{cs_check_key}
\alias{cs_check_token}
\title{Retrieve stored ChemSpider API keys and secure tokens}
\usage{
cs_check_key()

cs_check_token()
}
\value{
an API key
}
\description{
Look for and retrieve ChemSpider API keys and secure tokens stored in
.Renviron or .Rprofile.
}
\details{
To use the any of the functions in \code{webchem} that access the
ChemSpider database, you'll need to obtain an API key by registering.
Register at \url{https://developer.rsc.org/} for an API key. Please respect
the Terms & Conditions. The Terms & Conditions can be found at
\url{https://developer.rsc.org/terms}. Rather than store these API keys and
tokens in code, which might be shared, it's best practice to keep them
hidden. You can store the API key as \code{CHEMSPIDER_KEY=<your key>} in
.Renviron or as \code{option(chemspider_key = <your key>)} in .Rprofile.
Similarly, the secure token can be stored as \code{CHEMSPIDER_TOKEN} in
.Rprofile or as the option \code{chemspider_token}.
}
\examples{
\dontrun{
cs_check_key()
}
\dontrun{
cs_check_token()
}
}
\seealso{
\code{\link[usethis]{edit_r_environ}}
\code{\link[usethis]{edit_r_profile}}
}
