% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl.R
\name{chembl_atc_classes}
\alias{chembl_atc_classes}
\title{Retrieve all ATC classes}
\usage{
chembl_atc_classes(verbose = getOption("verbose"), test_service_down = FALSE)
}
\arguments{
\item{verbose}{logical; should a verbose output be printed on the console?}

\item{test_service_down}{logical; this argument is only used for testing.}
}
\description{
Retrieve all available classes within the Anatomical Therapeutic Chemical
(ATC) classification system.
}
\examples{
\dontrun{
# Might fail if API is not available

atc <- atc_classes()
}
}
\references{
Gaulton, A., Bellis, L. J., Bento, A. P., Chambers, J.,
Davies, M., Hersey, A., ... & Overington, J. P. (2012). ChEMBL: a large-scale
bioactivity database for drug discovery. Nucleic acids research, 40(D1),
D1100-D1107.
}
