% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl.R
\name{chembl_img}
\alias{chembl_img}
\title{Download images from ChEMBL}
\usage{
chembl_img(
  query,
  dir = NULL,
  verbose = getOption("verbose"),
  test_service_down = FALSE
)
}
\arguments{
\item{query}{character; a vector of ChEMBL IDs.}

\item{dir}{character; the directory in which to save the image(s). If the
directory does not exist, it will be created. If `NULL` (default), the
working directory will be used.}

\item{verbose}{logical; should verbose messages be printed to the console?}

\item{test_service_down}{logical; for internal testing only.}
}
\value{
Creates the download directory if needed and downloads one or more
SVG files.
}
\description{
Retrieve images of substances from ChEMBL and save them in SVG format.
}
\examples{
\dontrun{
# download a single image
chembl_img("CHEMBL1")
# download a single image to another directory
chembl_img("CHEMBL1", dir = tempdir())
# download multiple images
chembl_img(c("CHEMBL1", "CHEMBL2"))
}
}
