% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl.R
\name{chembl_resources}
\alias{chembl_resources}
\title{List ChEMBL Resources}
\usage{
chembl_resources()
}
\description{
Data in ChEMBL is organized in databases called resources. This function
lists available ChEMBL resources.
}
