% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl.R
\name{chembl_status}
\alias{chembl_status}
\title{Retrieve ChEMBL webservice status}
\usage{
chembl_status(verbose = getOption("verbose"), test_service_down = FALSE)
}
\arguments{
\item{verbose}{logical; should verbose messages be printed to the console?}

\item{test_service_down}{logical; for internal testing only.}
}
\description{
Retrieve status information about the ChEMBL webservice: database version,
release date, status, and count data for various entities.
}
