% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemid.R
\name{chemid}
\alias{chemid}
\title{Retrieve information from ChemIDPlus \url{http://chem.sis.nlm.nih.gov/chemidplus/name/triclosan}}
\usage{
chemid(query, type = c("rn", "name"), verbose = TRUE)
}
\arguments{
\item{query}{character; query string}

\item{type}{character; type of querystring.
'rn' for regeistry number or 'name' for common name.}
}
\value{
A list of 8 entries: name (vector), synonyms (vector), cas (vector),
inchi (vector), inchikey (vector), smiles(vector), toxicity (data.frame),
physprop (data.frame).
}
\description{
Retrieve information from ChemIDPlus \url{http://chem.sis.nlm.nih.gov/chemidplus/name/triclosan}
}
\note{
The data of the entry \code{physprop} is identical to the result returned
by \code{\link{physprop()}}.
}

