% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chooser}
\alias{chooser}
\title{Interactively choose a result from a menu}
\usage{
chooser(x, choices)
}
\arguments{
\item{x}{a character vector}

\item{choices}{If `choices = "all"` then the entire vector `x` is used for the menu.  If numeric > 1, only that number of elements from the start of `x` are shown. If `choices = 1`, then the first element of `x` is returned without prompting the user.  If `NULL` then `x` is returned unchanged.}
}
\value{
a character vector of length 1
}
\description{
In interactive sessions, prompts a user to choose an element of a vector from a menu. Use this for all functions that return multiple possible results such as multiple identifiers or synonyms.
}
\examples{
test <- c("apples", "bananas", "orange", "plum", "peach", "guava", "kumquat")
chooser(test, "all")
chooser(test, 3)
}
