% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{cs_convert_multiple}
\alias{cs_convert_multiple}
\title{Convert between SMILES, InChI, InChiKey, Mol identifiers using Chemspider}
\usage{
cs_convert_multiple(input, from, to, apikey = NULL)
}
\arguments{
\item{input}{character; the identifier to be converted.}

\item{from}{character; the format to be converted from. Valid values are
\code{"smiles"}, \code{"inchi"}, \code{"inchikey"}, \code{"mol"}.}

\item{to}{character; the format to be converted to. Valid values are
\code{"smiles"}, \code{"inchi"}, \code{"inchikey"}, \code{"mol"}.}

\item{apikey}{character; your API key. If NULL (default),
\code{cs_check_key()} will look for it in .Renviron or .Rprofile.}
}
\value{
Returns a character vector of length one containing the converted
identifier.
}
\description{
Submit a single identifier (SMILES, InChI, InChIKey or Mol) and return an
identifier in another format (SMILES, InChI, InChIKey or Mol). Not all
conversions are supported.
}
\details{
Not all conversions are supported. Allowed conversions:
\itemize{
\item InChI <-> InChIKey
\item InChI <-> SMILES
\item InChI -> Mol file
\item InChIKey <-> Mol file
}
}
\note{
An API key is needed. Register at \url{https://developer.rsc.org/}
for an API key. Please respect the Terms & Conditions. The Terms & Conditions
can be found at \url{https://developer.rsc.org/terms}.
}
\examples{
\dontrun{
cs_convert_multiple("CC(=O)O", "smiles", "inchi")
cs_convert_multiple("InChI=1S/C2H4O2/c1-2(3)4/h1H3,(H,3,4)", "inchi",
"inchikey")
cs_convert_multiple("QTBSBXVTEAMEQO-UHFFFAOYSA-N", "inchikey", "mol")
cs_convert_multiple("QTBSBXVTEAMEQO-UHFFFAOYSA-N", "inchikey", "mol",
parse = TRUE)
}
}
\references{
https://developer.rsc.org/compounds-v1/apis
}
\seealso{
This is a low level function and is not exported. See
\code{\link{cs_convert}} for the top level function.

\code{\link{parse_mol}}
}
\author{
Tamas Stirling, \email{stirling.tamas@gmail.com}
}
