% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{cs_element_csid}
\alias{cs_element_csid}
\title{ChemSpider ID from element specifications}
\usage{
cs_element_csid(includeElements, excludeElements, apikey,
  control = cs_control())
}
\arguments{
\item{includeElements}{character; a vector of up to 15 elements to include
in the query.}

\item{excludeElements}{character; a vector of up to 100 elements to exclude
from the query.}

\item{apikey}{character; your API key.}

\item{control}{function; see details.}
}
\value{
Returns a list of two elements.
}
\description{
Specify lists of elements to include and exclude from the query and retrieve
ChemSpider IDs (CSIDs) that match the given specifications.
}
\details{
See \code{cs_control()} for a full list of control options.

Control options available for this function are \code{orderBy},
\code{orderDirection}, \code{includeAll}, \code{complexity}, \code{isotopic}.
}
\note{
An API key is needed. Register at \url{https://developer.rsc.org/}
for an API key. Please respect the Terms & Conditions. The Terms & Conditions
can be found at \url{https://developer.rsc.org/terms}.
}
\references{
\url{https://developer.rsc.org/compounds-v1/apis}
}
\author{
Tamas Stirling, \email{stirling.tamas@gmail.com}
}
