% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{cs_formula_csid}
\alias{cs_formula_csid}
\title{Retrieve ChemSpider ID from compound formula}
\usage{
cs_formula_csid(formula, apikey = NULL, control = cs_control())
}
\arguments{
\item{formula}{character; search term.}

\item{apikey}{character; your API key. If NULL (default),
\code{cs_check_key()} will look for it in .Renviron or .Rprofile.}

\item{control}{list; see details.}
}
\value{
Returns a numeric vector of ChemSpider IDs.
}
\description{
Query a single compund by formula and return a vector of ChemSpider IDs.
}
\details{
\code{formula} can be expressed with and without LaTeX syntax.

Control options available for this function are \code{datasources},
\code{order_by}, \code{order_direction}. See \code{cs_control()} for a full
list of valid values for these control options.
}
\note{
An API key is needed. Register at RSC \url{https://developer.rsc.org/}
for an API key. Please respect the Terms & conditions
\url{https://developer.rsc.org/terms}.

This is a low level function and is not exported.
}
\examples{
\dontrun{
cs_formula_csid("C2H6O")
cs_formula_csid("C_{2}H_{6}O")
}
}
\references{
https://developer.rsc.org/compounds-v1/apis
}
\author{
Tamas Stirling, \email{stirling.tamas@gmail.com}
}
