% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{cs_inchi_csid}
\alias{cs_inchi_csid}
\title{Retrieve ChemSpider ID from InChI}
\usage{
cs_inchi_csid(inchi, apikey)
}
\arguments{
\item{inchi}{character; search term.}

\item{apikey}{character; your API key.}
}
\value{
Returns a list of two elements.
}
\description{
Retrieve ChemSpider ID from InChI
}
\note{
An API key is needed. Register at RSC
\url{https://developer.rsc.org/}
for an API key.
Please respect the Terms & conditions \url{https://developer.rsc.org/terms}.
}
\examples{
\dontrun{
apikey <- "<YOUR-API-KEY>"
cs_inchi_csid(inchi = "InChI=1S/C2H4O2/c1-2(3)4/h1H3,(H,3,4)",
apikey = apikey)
cs_inchi_csid(inchi = c(
"InChI=1S/C2H4O2/c1-2(3)4/h1H3,(H,3,4)",
"InChI=1/C2H4O/c1-2-3/h2H,1H3"), apikey = apikey)
}
}
\references{
https://developer.rsc.org/compounds-v1/apis
}
\seealso{
This is a low level function. See \code{\link{cs_convert}}
for the top level function.
}
\author{
Tamas Stirling, \email{stirling.tamas@gmail.com}
}
