% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{cs_name_csid}
\alias{cs_name_csid}
\title{Retrieve ChemSpider ID from compound name}
\usage{
cs_name_csid(name, apikey = NULL, control = cs_control())
}
\arguments{
\item{name}{character; search term.}

\item{apikey}{character; your API key. If NULL (default),
\code{cs_check_key()} will look for it in .Renviron or .Rprofile.}

\item{control}{list; see details.}
}
\value{
Returns a numeric vector of ChemSpider IDs.
}
\description{
Query a single compund by name and return a vector of ChemSpider IDs.
}
\details{
Control options available for this function are \code{order_by},
\code{order_direction}. See \code{cs_control()} for a full list of valid
values for these control options.
}
\note{
An API key is needed. Register at RSC \url{https://developer.rsc.org/}
for an API key. Please respect the Terms & conditions
\url{https://developer.rsc.org/terms}.

This is a low level function and is not exported.
}
\examples{
\dontrun{
cs_name_csid("ethanol")
}
}
\references{
https://developer.rsc.org/compounds-v1/apis
}
\author{
Tamas Stirling, \email{stirling.tamas@gmail.com}
}
