% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts.R
\name{cts_to}
\alias{cts_to}
\title{Return a list of all possible ids}
\usage{
cts_to(verbose = getOption("verbose"))
}
\arguments{
\item{verbose}{logical; should a verbose output be printed on the console?}
}
\value{
a character vector.
}
\description{
Return a list of all possible ids that can be used in the 'to' argument
}
\details{
See also \url{http://cts.fiehnlab.ucdavis.edu/}
}
\examples{
\dontrun{
cts_from()
}
}
\references{
Wohlgemuth, G., P. K. Haldiya, E. Willighagen, T. Kind, and O.
Fiehn 2010The Chemical Translation Service -- a Web-Based Tool to Improve
Standardization of Metabolomic Reports. Bioinformatics 26(20): 2647–2648.
}
\seealso{
\code{\link{cts_convert}}
}
