% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodb.R
\name{db_download_foodb}
\alias{db_download_foodb}
\title{Download FooDB database}
\usage{
db_download_foodb(verbose = getOption("verbose"))
}
\arguments{
\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
Invisibly returns the file path to the processed SQLite database.
}
\description{
Download the FooDB database for offline access.
}
\details{
FooDB is a static database with a single release. 
\code{db_download_foodb()} downloads the data set in JSON format and converts 
it to SQLite. If the database has already been downloaded and processed, the 
function will skip the download and conversion steps and simply return the 
file path to the existing database.
}
\note{
The SQLite database does not include spectra downloads. See
\url{https://foodb.ca/downloads} for more information on these data files.
}
\examples{
\dontrun{
db_download_foodb(verbose = TRUE)
}
}
