% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.format.cas}
\alias{.format.cas}
\title{Format a CAS number correctly}
\usage{
.format.cas(x)
}
\arguments{
\item{x}{a double or character to be converted to a CAS format}
}
\value{
correctly formatted CAS
}
\description{
Format a CAS number correctly
}
\author{
Eric Scott, \email{scottericr@gmail.com}
}
