% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ympa.R
\name{ec_query}
\alias{ec_query}
\title{Query Data bank of Environmental Properties of Chemicals (EnviChem)}
\usage{
ec_query(query, type = c("cas", "name", "free"), verbose = TRUE)
}
\arguments{
\item{query}{character; The searchterm}

\item{type}{character; Type of searchterm. Can be 'cas', 'name' or 'free' (for freetext)}

\item{verbose}{logical; print message during processing to console?}
}
\value{
a list with lists of 7 entries: name, cas, synonyms, formula, einecs,
molecular weigth and references
}
\description{
Query Data bank of Environmental Properties of Chemicals (EnviChem)
\url{https://wwwp.ymparisto.fi/scripts/kemrek/language.asp?Language=UK}
}
\examples{
\donttest{
# might fail if API is not available
h <- ec_query('50-00-0', 'cas')
h[['50-00-0']]$name

h <- ec_query(c('50-00-0', '3380-34-5'), 'cas')
}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

