% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etox.R
\name{etox_basic}
\alias{etox_basic}
\title{Get basic information from a ETOX ID}
\usage{
etox_basic(id, verbose = getOption("verbose"))
}
\arguments{
\item{id}{character; ETOX ID}

\item{verbose}{logical; print message during processing to console?}
}
\value{
a list with lists of four entries: cas (the CAS numbers), ec (the EC
  number), gsbl (the gsbl number), a data.frame synonys with synonyms and the
  source url.
}
\description{
Query ETOX: Information System Ecotoxicology and Environmental Quality
Targets \url{https://recherche.chemikalieninfo.de/etox} for basic information
}
\note{
Terms and Conditions: https://www.chemikalieninfo.de/termsofuse.html
}
\examples{
\dontrun{
id <- get_etoxid('Triclosan', match = 'best')
etox_basic(id$etoxid)

# Retrieve data for multiple inputs
ids <- c("20179", "9051")
out <- etox_basic(ids)
out

# extract cas numbers
sapply(out, function(y) y$cas)
}
}
\references{
Eduard Szöcs, Tamás Stirling, Eric R. Scott, Andreas Scharmüller,
Ralf B. Schäfer (2020). webchem: An R Package to Retrieve Chemical
Information from the Web. Journal of Statistical Software, 93(13).
\doi{10.18637/jss.v093.i13}.
}
\seealso{
\code{\link{get_etoxid}} to retrieve ETOX IDs,
  \code{\link{etox_basic}} for basic information, \code{\link{etox_targets}}
  for quality targets and \code{\link{etox_tests}} for test results
}
