% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etox.R
\name{etox_tests}
\alias{etox_tests}
\title{Get Tests from a ETOX ID}
\usage{
etox_tests(id, verbose = getOption("verbose"))
}
\arguments{
\item{id}{character; ETOX ID}

\item{verbose}{logical; print message during processing to console?}
}
\value{
A list of lists of two: A data.frame with test results from the ETOX database and the source_url.
}
\description{
Query ETOX: Information System Ecotoxicology and Environmental Quality Targets
\url{https://recherche.chemikalieninfo.de/etox} for tests
}
\note{
Terms and Conditions: https://www.chemikalieninfo.de/termsofuse.html
}
\examples{
\dontrun{
id <- get_etoxid('Triclosan', match = 'best')
out <- etox_tests(id$etoxid)
out[ , c('Organism', 'Effect', 'Duration', 'Time_Unit',
'Endpoint', 'Value', 'Unit')]
etox_tests( c("20179", "9051"))
}
}
\seealso{
\code{\link{get_etoxid}} to retrieve ETOX IDs, \code{\link{etox_basic}} for basic information,
\code{\link{etox_targets}} for quality targets and \code{\link{etox_tests}} for test results
}
