% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eup.R
\name{eup_list_entries}
\alias{eup_list_entries}
\title{List available entries in the local EU Pesticides database}
\usage{
eup_list_entries(idtype, verbose = getOption("verbose"))
}
\arguments{
\item{idtype}{character; the type of identifier to list. Allowed values are:
"substance_id", "substance_name", "as_cas_number", "pesticide_residue_id", 
"pesticide_residue_name".}

\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
A character vector of unique identifiers of the specified type that 
are present in the EU Pesticides database.
}
\description{
List available entries in the local EU Pesticides database
}
\examples{
\dontrun{
eup_list_entries("substance_name")
eup_list_entries("pesticide_residue_name")
}
}
\references{
You can find more information about the EU Pesticides database at
\url{https://food.ec.europa.eu/plants/pesticides/eu-pesticides-database_en}.
}
