% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eup.R
\name{eup_query}
\alias{eup_query}
\title{Query EU Pesticides}
\usage{
eup_query(query, resource, mode = "offline", ...)
}
\arguments{
\item{query}{numeric; a vector of IDs. The type of ID depends on the
resource. See examples for more information.}

\item{resource}{character; the EU Pesticides resource to query. Can be one of
\code{"active_substances"} or \code{"residues"}.}

\item{mode}{character; the mode of operation. Can be one of \code{"offline"} 
(offline access) or \code{"ws"} (web service access). Currently only offline 
mode is implemented.}

\item{...}{Further args passed on to [DBI::dbConnect()]}
}
\value{
A data frame containing information about the specified entry.
}
\description{
Query EU Pesticides
}
\examples{
\dontrun{
# Download database
db_download_eup(verbose = TRUE)

# Retrieve information about active substances
eup_query(query = c(1313, 1314), resource = "active_substances")

# Retrieve information about residues
eup_query(query = c(1, 2), resource = "residues")
}
}
\references{
You can find more information about the EU Pesticides database at
\url{https://food.ec.europa.eu/plants/pesticides/eu-pesticides-database_en}.
}
