% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extr_num}
\alias{extr_num}
\title{Extract number from string}
\usage{
extr_num(x)
}
\arguments{
\item{x}{character; input string}
}
\value{
a numeric vector
}
\description{
Extract number from string
}
\examples{
extr_num('aaaa -95')
}

