% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractors}
\alias{cas}
\alias{extractors}
\title{Extract parts from webchem objects}
\usage{
cas(...)
}
\arguments{
\item{x}{object}
}
\value{
a vector of cas numbers
}
\description{
Extract parts from webchem objects
}

