% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodb.R
\name{foodb_convert}
\alias{foodb_convert}
\title{Convert compound identifiers in the local FooDB database}
\usage{
foodb_convert(query, from, to, verbose = getOption("verbose"))
}
\arguments{
\item{query}{character; a character of compound identifiers to convert.}

\item{from}{The type of identifier provided in \code{query}. See Details for
allowed values.}

\item{to}{The type of identifier to convert to. See Details for allowed
values.}

\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
A character vector of converted identifiers, in the same order as
the input \code{query}. If an identifier could not be converted, the
corresponding output will be \code{NA}.
}
\description{
Convert compound identifiers in the local FooDB database
}
\details{
Allowed values for \code{from} and \code{to} are:
\itemize{
  \item \code{id}: internal numeric ID in the FooDB database
  \item \code{public_id}: public numeric ID in the FooDB database
  \item \code{name}: compound name
  \item \code{cas_number}: CAS number
  \item \code{moldb_smiles}: SMILES string
  \item \code{moldb_inchi}: InChI string
  \item \code{moldb_inchikey}: InChIKey string
  \item \code{moldb_iupac}: IUPAC name
}
}
\examples{
\dontrun{
foodb_convert("4", from = "id", to = "name")
foodb_convert("FDB000004", from = "public_id", to = "cas_number")
}
}
