% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodb.R
\name{foodb_list_compounds}
\alias{foodb_list_compounds}
\title{List available compound identifiers in the local FooDB database}
\usage{
foodb_list_compounds(
  idtype,
  include_synonyms = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{idtype}{character; the type of identifier to list. Allowed values are:
"id", "public_id", "name", "cas_number", "moldb_smiles", "moldb_inchi", 
"moldb_inchikey", "moldb_iupac".}

\item{include_synonyms}{logical; should compound name synonyms also be 
included in the output? Only used if `idtype` is "name". Default is `TRUE`.}

\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
A character vector of unique identifiers of the specified type that 
are present in the FooDB database.
}
\description{
List available compound identifiers in the local FooDB database
}
\examples{
\dontrun{
foodb_list_compounds("name")
foodb_list_compounds("cas_number")
}
}
