% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodb.R
\name{foodb_query}
\alias{foodb_query}
\title{Query the local FooDB database for compound information}
\usage{
foodb_query(query, from, verbose = getOption("verbose"))
}
\arguments{
\item{query}{character; a vector of compound identifiers to query.}

\item{from}{character; the type of identifier provided in `query`. Allowed 
values are: "id", "public_id", "name", "cas_number", "moldb_smiles", 
"moldb_inchi", "moldb_inchikey", "moldb_iupac".}

\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
A nested list of fields for each compound query, with available 
information from the FooDB database. If a query does not match any compound 
in the database, the corresponding output will be `NA`.
}
\description{
Query the local FooDB database for compound information
}
\examples{
\dontrun{
# single query
foodb_query("Biotin", from = "name")
foodb_query("YBJHBAHKTGYVGT-UHFFFAOYSA-N", from = "moldb_inchikey")

# multiple queries
foodb_query(c("Biotin", "Folic acid"), from = "name")
}
}
