% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugbank.R
\name{get_dbid}
\alias{get_dbid}
\title{Get Drugbank ID}
\usage{
get_dbid(query, match = c("best", "all", "first", "ask", "na"),
  verbose = TRUE)
}
\arguments{
\item{query}{character; The searchterm}

\item{match}{character; How should multiple hits be handeled? 'all' returns all matched IDs,
'first' only the first match, 'best' the best matching (by name) ID, 'ask' is a interactive mode and the user is asked for input,
'na' returns NA if multiple hits are found.}

\item{verbose}{logical; print message during processing to console?}
}
\value{
if match = 'all' a list with etoxids, otherwise a dataframe with 4 columns:
etoxID, matched substance, string distance to match and the queried string
}
\description{
Query Drugbank
\url{http://www.drugbank.ca/} for their Drugbank ID
}
\note{
Before using this function, please read the disclaimer \url{http://www.drugbank.ca/about}.
}
\examples{
\dontrun{
# might fail if API is not available
get_dbid('Triclosan', match = 'all')
get_dbid('Atrazin', match = 'all')
get_dbid('Atrazin', match = 'best')
get_dbid('Atrazin', match = 'na')
get_dbid('Atrazin', match = 'first')
get_dbid('Atrazin', match = 'ask')
get_dbid(c('Triclosan', 'Atrazin'), match = 'best')
}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}
\seealso{

}

