% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST.R
\name{get_RI}
\alias{get_RI}
\title{Retrieve retention indices from NIST}
\usage{
get_RI(cas, type = c("kovats", "linear", "alkane", "lee"),
  polarity = c("polar", "non-polar"), temp_prog = c("isothermal", "ramp",
  "custom"))
}
\arguments{
\item{cas}{CAS numbers either as numeric or formatted correctly with hyphens.}

\item{type}{Type of RI to retrieve.  Details about how these are calculated here: [https://webbook.nist.gov/chemistry/gc-ri/]}

\item{polarity}{"polar" to get RIs for polar columns, "non-polar" for non-polar columns}

\item{temp_prog}{one of "isothermal", "ramp", or "custom"}
}
\value{
a table of literature RIs
}
\description{
This function scrapes NIST for literature retention indices given CAS numbers as an input.
}
\examples{
myRIs <- get_RI(c("78-70-6", "873-94-9", "13474-59-4"), "linear", "non-polar", "ramp")
}
\seealso{
\code{\link{is.cas}} \code{\link{as.cas}}
}
