% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST.R
\name{get_RI_xml}
\alias{get_RI_xml}
\title{Scrape Retention Indices from NIST}
\usage{
get_RI_xml(cas, type = c("kovats", "linear", "alkane", "lee"),
  polarity = c("polar", "non-polar"), temp_prog = c("isothermal", "ramp",
  "custom"))
}
\arguments{
\item{cas}{cas number}

\item{type}{what kind of RI}

\item{polarity}{polar or non-polar}

\item{temp_prog}{what kind of temperature program}
}
\value{
an xml nodeset
}
\description{
Scrape Retention Indices from NIST
}
