% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{get_sid}
\alias{get_sid}
\title{Retrieve PubChem Substance ID (SID)}
\usage{
get_sid(
  query,
  from = "name",
  match = c("all", "first", "ask", "na"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{query}{character; search term, one ore more substances or depositors.}

\item{from}{character; type of input. Valid values are \code{"name"},
\code{<xref>}, \code{"sourceid/<source id>"} or \code{"sourceall"}. See
details for more information.}

\item{match}{character; How should multiple hits be handled?, "all" all
matches are returned, "best" the best matching is returned, "ask" enters an
interactive mode and the user is asked for input, "na" returns NA if multiple
hits are found.}

\item{verbose}{logical; should a verbose output be printed on the console?}

\item{...}{currently not used.
#' @details \code{<xref>} is assembled as "\code{xref}/\{\code{RegistryID} |
\code{RN} | \code{PubMedID} | \code{MMDBID} | \code{ProteinGI},
\code{NucleotideGI} | \code{TaxonomyID} | \code{MIMID} | \code{GeneID} |
\code{ProbeID} | \code{PatentID}\}", e.g. \code{from = "xref/RN"} will query by
CAS RN.}
}
\value{
Returns a tibble of substance id-s.
}
\description{
Retrieve substance IDs (SIDs) from PubChem based on substance
name, registry ID (e.g. CAS RN), or source ID. Alternatively, retrieve SIDs
of all substances provided by a PubChem depositor.
}
\details{
\code{<source id>} is any valid PubChem Data Source ID.

If \code{from = "sourceall"} the query is one or more valid Pubchem
depositor names. Depositor names are not case sensitive, but are sensitive to
spaces, and special characters may need to be escaped, such as "&" should be
replaced by "\%26".

Depositor names and Data Source IDs can be found at
\url{https://pubchem.ncbi.nlm.nih.gov/sources/}.
}
\note{
Please respect the Terms and Conditions of the National Library of
Medicine, \url{https://www.nlm.nih.gov/databases/download.html} and the data
usage policies of National Center for Biotechnology Information,
\url{https://www.ncbi.nlm.nih.gov/home/about/policies/},
\url{https://pubchemdocs.ncbi.nlm.nih.gov/programmatic-access}.
}
\examples{
# might fail if API is not available
\donttest{
get_sid("2-(Acetyloxy)benzoic acid", from = "name")
get_sid("57-27-2", from = "xref/rn")
get_sid(c("VCC957895", "VCC883692"), from = "sourceid/23706")
get_sid("Optopharma Ltd", from = "sourceall")
}
}
\references{
Wang, Y., J. Xiao, T. O. Suzek, et al. (2009). PubChem: A Public
Information System for Analyzing Bioactivities of Small Molecules. Nucleic
Acids Research 37: 623–633.

Kim, S., Thiessen, P. A., Bolton, E. E., & Bryant, S. H. (2015).
PUG-SOAP and PUG-REST: web services for programmatic access to chemical
information in PubChem. Nucleic acids research, gkv396.

Kim, Sunghwan, Paul A. Thiessen, Evan E. Bolton, et al. (2016).
PubChem Substance and Compound Databases. Nucleic Acids Research 44(D1):
D1202–D1213.

Sunghwan Kim, Paul A Thiessen, Tiejun Cheng, Bo Yu, Evan E Bolton
(2018) An update on PUG-REST: RESTful interface for programmatic access to
PubChem. Nucleic Acids Research 46(W1): W563–W570.
}
\author{
Tamás Stirling, \email{stirling.tamas@gmail.com}
}
