% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{get_wdid}
\alias{get_wdid}
\title{Get Wikidata Item ID}
\usage{
get_wdid(query, language, match = c("all", "first", "best"), verbose = TRUE)
}
\arguments{
\item{query}{character; The searchterm}

\item{language}{character; the language to search in}

\item{match}{character; \code{match = "all"} returns all matches,
\code{match = "first"} the first one and \code{match = "best"} the hit with the lowest
 Levenshtein distance between query and hit.}

\item{verbose}{logical; print message during processing to console?}
}
\value{
A character vector with the item ID and additional attributes \code{matched}  (the matched
substance name) and \code{distance} (the normalized string distance of the query to the match).
}
\description{
Get Wikidata Item ID
}
\note{
Only matches in labels are returned.
}
\examples{
\dontrun{
get_wdid('Triclosan', language = 'en')
# multiple inpus
comps <- c('Triclosan', 'Glyphosate')
sapply(comps, get_wdid, language = 'en')
}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

