% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.inchikey}
\alias{is.inchikey}
\title{Check if input is a valid inchikey}
\usage{
is.inchikey(x)
}
\arguments{
\item{x}{character; input string}
}
\value{
a logical
}
\description{
Check if input is a valid inchikey
}
\examples{
is.inchikey('BQJCRHHNABKAKU-KBQPJGBKSA-N')
is.inchikey('BQJCRHHNABKAKU-KBQPJGBKS')
}

