% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webchem.R
\docType{data}
\name{jagst}
\alias{jagst}
\title{Organic plant protection products in the river Jagst / Germany in 2013}
\format{A data frame with 442 rows and 4 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
}}
\source{
\url{http://jdkfg.lubw.baden-wuerttemberg.de/servlet/is/300/}
}
\usage{
jagst
}
\description{
This dataset comprises environmental monitoring data of organic plant protection products
in the year 2013 in the river Jagst, Germany.
The data is publicly available and can be retrieved from the
LUBW Landesanstalt für Umwelt, Messungen und Naturschutz Baden-Württemberg.
It comprises measurement 34 substances (substance without detects have been removed) on 13 sampling occasions.
}
\keyword{datasets}

