% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chebi.R
\name{l2df}
\alias{l2df}
\title{Helper function to parse some ChEBI data}
\usage{
l2df(x)
}
\arguments{
\item{x}{list; a list to bind into a data.frame}
}
\value{
a data.frame
}
\description{
Helper function to parse some ChEBI data
}
\seealso{
\code{\link{chebi_comp_entity}}
}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
