% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-test.R
\name{N}
\alias{N}
\title{Testing testing \code{\link{testing}}}
\usage{
N()
}
\value{
an error
}
\description{
This should break the pkgdown build (https://github.com/r-lib/pkgdown/issues/1351)
}
\examples{
N()
}
