% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{pc_page}
\alias{pc_page}
\title{Import PubChem content pages}
\usage{
pc_page(cid, section, verbose = TRUE)
}
\arguments{
\item{cid}{numeric; a vector of identifiers to search for.}

\item{section}{character; the section of the conent page to be imported.}
}
\value{
a list of PubChem content pages.
}
\description{
When you search for a compound at \url{https://pubchem.ncbi.nlm.nih.gov/},
and select the compound you are interested in, you will be forwarded to a
PubChem content page. This function is used to import a section of that page
into your workflow.
}
\details{
\code{section} can be any section of a PubChem content page, e.g.
\code{section = "solubility"} will import the section on solubility, or
\code{section = "experimental properties"} will import all experimental
properties. The \code{section} argument is not case sensitive but it
is sensitive to typing errors and it requires the full name of the section as
it is printed on the content page. The PubChem Table of Contents Tree can
also be found at
\url{https://pubchem.ncbi.nlm.nih.gov/classification/#hid=72}.

The function will return a list of content pages where each
element is a list itself. While it is possible to navigate the returned
objects through list operations, it is more convenient to use
\code{pc_extract()} to access low level data.
}
\note{
Please respect the Terms and Conditions of the National Library of
Medicine, \url{https://www.nlm.nih.gov/databases/download.html} the data
usage policies of National Center for Biotechnology Information,
\url{https://www.ncbi.nlm.nih.gov/home/about/policies/},
\url{https://pubchemdocs.ncbi.nlm.nih.gov/programmatic-access}, and the data
usage policies of the indicidual data sources
\url{https://pubchem.ncbi.nlm.nih.gov/sources/}.
}
\examples{
\donttest{
pka <- pc_page(176, "pKa")
ep <- pc_page(c(176, 311), "experimental properties")
tox <- pc_page(176, "toxicity")
}
}
\author{
Tamas Stirling, \email{stirling.tamas@gmail.com}
}
