% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{pc_pugrest}
\alias{pc_pugrest}
\title{Search PubChem through highly customizable requests}
\usage{
pc_pugrest(query, domain, from, to = "all", output)
}
\arguments{
\item{query}{character; search term. A comma-separated list of positive
integers (e.g. cid, sid, aid) or identifier strings (source, inchikey,
formula). In some cases only a single identifier string (name, smiles, xref;
inchi, sdf by POST only).}

\item{domain}{character; the domain of the search term. Can be one of
\code{"compound"}, \code{"substance"}, \code{"assay"},
\code{"sources/[substance, assay]"}, \code{"sourcetable"},
\code{"conformers"},
\code{"annotations/[sourcename/<source name>|heading/<heading>]"}. See
details for more information.}

\item{from}{character; the type of the search term, refered to as
\code{namespace} within the reference. Valid values depend on the domain.
See details for more information.}

\item{to}{character; refered to as \code{operations} within the reference.
Tells the service which part of the record to retrieve. Currently, if no
operation is specified at all, the default is to retrieve the entire record.
See details for more information.}

\item{output}{character; the format of the output. Can be one of
\code{"XML"}, \code{"ASNT"}, \code{"ASNB"}, \code{"JSON"},
\code{"JSONP[?callback=<callback name>]"}, \code{"SDF"}, \code{"CSV"},
\code{"PNG"}, \code{"TXT"}. Note that not all formats are applicable to the
results of all operations; one cannot, for example, retrieve a whole compound
record as CSV or a property table as SDF. TXT output is only available in a
restricted set of cases where all the information is the same – for example,
synonyms for a single CID where there is one synonym per line.}
}
\value{
a html POST response.
}
\description{
This function gives you more control over your communications
with the PubChem PUG-REST webservice. This documentation page was created
from the original PubChem PUG-REST specification document which can be found
at \url{https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest}.
}
\details{
Valid \code{from} values for \code{domain = "compound"} are:
\itemize{
\item \code{"cid"}, \code{"name"}, \code{"smiles"}, \code{"inchi"},
\code{"sdf"}, \code{"inchikey"}, \code{"formula"}, \code{<structure search>},
\code{<xref>}, \code{"listkey"}, \code{<fast search>}.
\item \code{<structure search>} is assembled as "\{\code{substructure},
\code{superstructure}, \code{similarity}, \code{identity}\}/\{\code{smiles},
\code{inchi}, \code{sdf}, \code{cid}\}", e.g.
\code{from = "substructure/smiles"}.
\item \code{<xref>} is assembled as "\code{xref}/\{\code{RegistryID},
\code{RN}, \code{PubMedID}, \code{MMDBID}, \code{ProteinGI},
\code{NucleotideGI}, \code{TaxonomyID}, \code{MIMID}, \code{GeneID},
\code{ProbeID}, \code{PatentID}\}, e.g. \code{from = "xref/RN"}. <xref> is
not case sensitive.
\item \code{<fast search>} is either \code{"fastformula"} or is assembled as
"\{\code{fastidentity}, \code{fastsimilarity_2d}, \code{fastsimilarity_3d},
\code{fastsubstructure}, \code{fastsuperstructure}\}/\{\code{smiles},
\code{smarts}, \code{inchi}, \code{sdf} or \code{cid}\}", e.g.
\code{from = "fastidentity/smiles"}.
}

Valid \code{from} values for \code{domain = "substance"} are:
\itemize{
\item \code{"sid"}, \code{"sourceid/<source id>"},
\code{"sourceall/<source name>"}, \code{name}, \code{<xref>} or
\code{listkey}.
\item \code{<source name>} is any valid PubChem depositor name.
}

Valid \code{from} values for \code{domain = "assay"} are:
\itemize{
\item \code{"aid"}, \code{"listkey"}, \code{"type/<assay type>"},
\code{"sourceall/<source name>"}, \code{"target/<assay target>"} or
\code{"activity/<activity column name>"}.
\item \code{<assay type>} can be \code{all}, \code{confirmatory},
\code{doseresponse}, \code{onhold}, \code{panel}, \code{rnai},
\code{screening}, \code{summary}, \code{cellbased}, \code{biochemical},
\code{invivo}, \code{invitro} or \code{activeconcentrationspecified}.
\item \code{<assay target>} can be \code{gi}, \code{proteinname},
\code{geneid}, \code{genesymbol} or \code{accession}.
}

\code{<other inputs>} for domain can be \code{"sources/[substance,
assay]"}, \code{"sourcetable"}, \code{"conformers"}, \code{"annotations/
[sourcename/<source name>, heading/<heading>]"}.

Valid \code{to} values for \code{domain = "compound"} are:
\itemize{
\item \code{"record"}, \code{<compound property>}, \code{"synonyms"},
\code{"sids"}, \code{"cids"}, \code{"aids"}, \code{"assaysummary"},
\code{"classification"}, \code{<xrefs>}, \code{"description"} or
\code{"conformers"}.
\item \code{<compound property>} is assembled as \code{"property/
[comma-separated list of <property tags>]"}.
\item Valid \code{<property tags>} are \code{"MolecularFormula"},
\code{"MolecularWeight"}, \code{"CanonicalSMILES"}, \code{"IsomericSMILES"},
\code{"InChI"}, \code{"InChIKey"}, \code{"IUPACname"}, \code{"XlogP"},
\code{"ExactMass"}, \code{"MonoisotopicMass"}, \code{"TPSA"},
\code{"Complexity"}, \code{"Charge"}, \code{"HBondDonorCount"},
\code{"HBondAcceptorCount"}, \code{"RotatableBondCount"},
\code{"HeavyAtomCount"}, \code{"IsotopeAtomCount"}, \code{"AtomStereoCount"},
\code{"DefinedAtomStereoCount"}, \code{"UndefinedAtomStereoCount"},
\code{"BondStereoCount"}, \code{"DefinedBondStereoCount"},
\code{"UndefinedBondStereoCount"}, \code{"CovalentUnitCount"},
\code{"Volume3D"}, \code{"XStericQuadrupole3D"},
\code{"YStericQuadrupole3D"}, \code{"ZStericQuadrupole3D"},
\code{"FeatureCount3D"}, \code{"FeatureAcceptorCount3D"},
\code{"FeatureDonorCount3D"}, \code{"FeatureAnionCount3D"},
\code{"FeatureCationCount3D"}, \code{"FeatureRingCount3D"},
\code{"FeatureHydrophobeCount3D"}, \code{"ConformerModelRMSD3D"},
\code{"EffectiveRotorCount3D"}, \code{"ConformerCount3D"},
\code{"Fingerprint2D"}. <property tags> are not case sensitive.
}

Valid \code{to} values for \code{domain = "substance"} are:
\itemize{
\item \code{"record"}, \code{"synonyms"}, \code{"sids"}, \code{"cids"},
\code{"aids"}, \code{"assaysummary"}, \code{"classification"},
\code{<xrefs>}, \code{"description"}.
\item \code{<xrefs>} is assembled as \code{"xrefs/
[comma-separated list of xrefs tags]"}.
}

Valid \code{to} values for \code{domain = "assay"} are:
\itemize{
\item \code{"record"}, \code{"concise"}, \code{"aids"}, \code{"sids"},
\code{"cids"}, \code{"description"}, \code{"targets/<target type>"},
\code{<doseresponse>}, \code{"summary"} or \code{"classification"}.
\item \code{<target type>}  can be \code{"ProteinGI"}, \code{"ProteinName"},
\code{"GeneID"} or \code{"GeneSymbol"}.
\item \code{<doseresponse>} is assembled as \code{"doseresponse/sid"}.
}
}
\note{
Other special characters may need to be escaped, such as "&" should be
replaced by "\%26".
}
\references{
For more information, visit
\url{https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest}
}
\author{
Tamás Stirling, \email{stirling.tamas@gmail.com}
}
