% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{pc_query}
\alias{pc_query}
\title{Query PubChem for custom information}
\usage{
pc_query(cid, query)
}
\description{
Query PubChem for custom information
}
\details{
The full list of queries can be found at
\url{https://pubchem.ncbi.nlm.nih.gov/classification/#hid=72}
}
\examples{
\donttest{
pc_query(176, "pka")

}

}
