% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{ping}
\alias{ping_pubchem}
\alias{ping_pubchem_pw}
\alias{ping_cs}
\alias{ping_pan}
\alias{ping_aw}
\alias{ping_chebi}
\alias{ping_ci}
\alias{ping_cir}
\alias{ping_cts}
\alias{ping_etox}
\alias{ping_fn}
\alias{ping_nist}
\alias{ping_opsin}
\alias{ping_wd}
\title{Ping an API used in webchem to see if it's working.}
\usage{
ping_pubchem(...)

ping_pubchem_pw(...)

ping_cs(...)

ping_pan(...)

ping_aw(...)

ping_chebi(...)

ping_ci(...)

ping_cir(...)

ping_cts(...)

ping_etox(...)

ping_fn(...)

ping_nist(...)

ping_opsin(...)

ping_wd(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}
}
\value{
A logical, TRUE or FALSE

TRUE if pubchem is reachable

TRUE if pubchem PUG-VIEW is reachable

TRUE if ChemSpider is reachable

TRUE if PAN is reachable

TRUE if www.alanwood.net/pesticides is reachable

TRUE if ChEBI is reachable

TRUE if ChemIDPlus is reachable

TRUE if Chemical Identity Resolver is reachable

TRUE if Chemical Translation Service is reachable

TRUE if ETOX is reachable

TRUE if flavornet.org is reachable

TRUE if NIST Web Book is reachable

TRUE if OPSIN is reachable

TRUE if Wikidata is reachable
}
\description{
Ping an API used in webchem to see if it's working.
}
\examples{
\dontrun{
 # might fail if API is not available
 ping_pubchem()
 }
\dontrun{
 # might fail if API is not available
 ping_pubchem_pw()
 }
\dontrun{
 # might fail if API is not available
 ping_cs()
 }
\dontrun{
 # might fail if API is not available
 ping_pan()
 }
\dontrun{
 # might fail if site is not available
 ping_aw()
 }
\dontrun{
 # might fail if site is not available
 ping_chebi()
 }
\dontrun{
 # might fail if site is not available
 ping_ci()
 }
\dontrun{
 # might fail if site is not available
 ping_cir()
 }
\dontrun{
 # might fail if site is not available
 ping_cts()
 }
\dontrun{
 # might fail if site is not available
 ping_etox()
 }
\dontrun{
 # might fail if site is not available
 ping_fn()
 }
\dontrun{
 # might fail if site is not available
 ping_nist()
 }
\dontrun{
 # might fail if site is not available
 ping_opsin()
 }
\dontrun{
 # might fail if site is not available
 ping_wd()
 }
}
