% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.R
\name{ping_service}
\alias{ping_service}
\title{Ping an API used in webchem to see if it's working.}
\usage{
ping_service(
  service = c("bcpc", "chebi", "ci", "cs", "cs_web", "cir", "cts", "etox", "fn",
    "nist", "opsin", "pan", "pc", "srs", "wd")
)
}
\arguments{
\item{service}{character; the same abbreviations used as prefixes in \code{webchem} functions, with the exception of \code{"cs_web"}, which only checks if the ChemSpider website is up, and thus doesn't require an API key.}
}
\value{
A logical, TRUE if the service is available or FALSE if it isn't
}
\description{
Ping an API used in webchem to see if it's working.
}
\examples{
\dontrun{
ping_service("pan")
}
}
