% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ppdb.R
\docType{data}
\name{ppdb_idx}
\alias{ppdb_idx}
\title{Sappschot of the PPDB search index}
\format{A data frame with 27383 rows and 2 variables:
\describe{
  \item{match}{string}
  \item{link}{matched link}
}}
\source{
\url{http://sitem.herts.ac.uk/aeru/iupac/search.htm}
}
\usage{
ppdb_idx
}
\description{
A dataset containing matched strings and links to the PPDB \url{http://sitem.herts.ac.uk/aeru/iupac/search.htm}
Build using \code{\link{ppdb_buildidx}} on 23th August 2015.
}
\seealso{
\code{\link{ppdb_buildidx}}
}
\keyword{datasets}

