% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ppdb.R
\name{ppdb_query}
\alias{ppdb_query}
\title{Query the ppdb for information}
\usage{
ppdb_query(cas, verbose = TRUE)
}
\arguments{
\item{verbose}{logical; print message during processing to console?}

\item{character;}{CAS number to query.}
}
\value{
A list of 9 data.frames: ec_regulation, approved_in, general, fate,
deg, soil, metab, etox and names.




See also \url{http://sitem.herts.ac.uk/aeru/iupac/docs/Background_and_Support.pdf} for more information on the data
}
\description{
This function queries the PPDB \url{http://sitem.herts.ac.uk/aeru/iupac/search.htm} for information.
}
\examples{
\dontrun{
ppdb_query('1071-83-6')
}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

