% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chebi.R
\name{rbind.named.fill}
\alias{rbind.named.fill}
\title{Helper function replacing do.call(rbind, list())
to address the issue of different column lengths in a list of data.frames
taken from: https://stackoverflow.com/questions/17308551/do-callrbind-list-for-uneven-number-of-column}
\usage{
\method{rbind}{named.fill}(x)
}
\arguments{
\item{x}{list; a list to bind into a data.frame}
}
\description{
Helper function replacing do.call(rbind, list())
to address the issue of different column lengths in a list of data.frames
taken from: https://stackoverflow.com/questions/17308551/do-callrbind-list-for-uneven-number-of-column
}
\seealso{
\code{\link{l2df}}
}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
