% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST.R
\name{read_html_slow}
\alias{read_html_slow}
\title{Read HTML slowly}
\usage{
read_html_slow(x, ...)
}
\arguments{
\item{x}{a URL}

\item{...}{currently unused}
}
\value{
html
}
\description{
Just adds a rest after `read_html()`.  Useful for web-scraping.
}
