% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smarts_viewer.R
\name{smarts_viewer}
\alias{smarts_viewer}
\title{Query SMARTS Viewer}
\usage{
smarts_viewer(
  smarts,
  output = c("image", "download"),
  image_format = c("pdf", "png", "svg"),
  visualization_modus = c(1, 2),
  legend_option = c("both", "none", "static", "dynamic"),
  filename = NULL
)
}
\arguments{
\item{smarts}{A SMARTS string to visualize}

\item{output}{The type of response generated (image or download). See
description for details.}

\item{image_format}{Image file format (pdf, png, or svg)}

\item{visualization_modus}{1 or 2 (1 = Complete Visualization, 2 = Element
Symbols)}

\item{legend_option}{both, none, static, dynamic}

\item{filename}{Filename to use if output is to be downloaded}
}
\value{
Either a raster object (output = image) suitable for adding to an
  R plot using \code{rasterImage} or the image is downloaded in the specified
  format to the indicated filename.
}
\description{
Submit a query to the SMARTS viewer webservice at
\url{http://smartsview.zbh.uni-hamburg.de}, ZBH Center for Bioinformatics,
University of Hamburg
}
\examples{
\dontrun{
img <- smarts_viewer(
   "[CX3](=[OX1])[OX2][CX3](=[OX1])",
   "image", "png", 1, "both"
   )
plot(0:1,0:1, 'n')
rasterImage(img[[1]],0,0,1,1)

smarts <- c("[CX3](=[OX1])[OX2][CX3](=[OX1])",
"[$([nr5]:[nr5,or5,sr5]),$([nr5]:[cr5]:[nr5,or5,sr5])]",
"[#6][$([NX2]=O),$(N=C=O),$(OC#N),$(SC#N)]"
)
img <- smarts_viewer(smarts, "image", "png", 1, "both")
par(mfcol = c(1,length(smarts)))
sapply(img, function(i){
 plot(0:1,0:1, 'n')
 rasterImage(i,0,0,1,1)
})

smarts_viewer(
    smarts, output = "download",
    image_format = "pdf",
    visualization_modus = 1,
    legend_option = "both", filename = "test.pdf")
}

}
