% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodscents.R
\name{tgsc_percept}
\alias{tgsc_percept}
\title{Scrape \url{thegoodsscentscompany.com} for aromas of compounds}
\usage{
tgsc_percept(CAS, flavor = FALSE)
}
\arguments{
\item{CAS}{character; CAS number to search by. See \code{\link{is.cas}} for
correct formatting.}

\item{flavor}{boolean; FALSE by default, if TRUE, returns flavor instead of
odor}
}
\value{
Returns a character value of the found aromas or an NA if not found.
}
\description{
Scrape \url{thegoodsscentscompany.com} for aromas of compounds
}
\examples{
\donttest{
tgsc_percept('66-25-1')
}
}
\author{
Josh Morimoto, \email{joshua.morimoto@tufts.edu}
}
