% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST.R
\name{tidy_RItable}
\alias{tidy_RItable}
\title{Tidier for webscraped RI RI_xml}
\usage{
tidy_RItable(RI_xml)
}
\arguments{
\item{RI_xml}{captured by `get_RI_xml`}
}
\value{
a single table
}
\description{
Tidier for webscraped RI RI_xml
}
