% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webchem-deprecated.R
\name{webchem-deprecated}
\alias{cid_compinfo}
\alias{get_cid}
\alias{webchem-deprecated}
\title{Deprecated function(s) in the webchem package}
\usage{
get_cid(...)

cid_compinfo(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the webchem package.  They may eventually be completely
removed.
}
\details{
Deprecated functions are:
\tabular{rl}{
  \code{get_pcid} \tab is now a synonym for \code{\link{get_cid}}\cr
  \code{pc_compinfo} \tab is now a synonym for \code{\link{cid_compinfo}}\cr
}
}

