% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cas}
\alias{cas}
\alias{cas.default}
\title{Extract CAS numbers from a webchem object}
\usage{
cas(...)

\method{cas}{default}(x)
}
\arguments{
\item{x}{object}
}
\value{
a vector of cas numbers
}
\description{
Extract CAS numbers from a webchem object
}

