% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webchem-package.R
\docType{package}
\name{webchem-package}
\alias{webchem}
\alias{webchem-package}
\title{webchem: An R package to retrieve chemical information from the web.}
\description{
Chemical information from around the web. This package interacts with a suite
of web APIs for chemical information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/webchem/}
  \item \url{https://github.com/ropensci/webchem}
  \item Report bugs at \url{https://github.com/ropensci/webchem/issues}
}

}
\author{
\strong{Maintainer}: Tamás Stirling \email{stirling.tamas@gmail.com} [contributor]

Authors:
\itemize{
  \item Eduard Szöcs
}

Other contributors:
\itemize{
  \item Robert Allaway [contributor]
  \item Daniel Muench [contributor]
  \item Johannes Ranke [contributor]
  \item Andreas Scharmüller [contributor]
  \item Eric R Scott [contributor]
  \item Jan Stanstrup [contributor]
  \item João Vitor F Cavalcante [contributor]
  \item Gordon Getzinger [contributor]
  \item Ethan Bass [contributor]
}

}
