% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webchem-package.R
\docType{package}
\name{webchem}
\alias{webchem}
\title{webchem: An R package to retrieve chemical information from the web.}
\description{
Chemical information from around the web. This package interacts with a suite
of web APIs for chemical information.
}
