% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comptox.R
\name{webtest_query}
\alias{webtest_query}
\title{Retrieve data from the Web-services Toxicity Estimation Software Tool (WebTEST)}
\usage{
webtest_query(
  query,
  from = "smiles",
  endpoint = c("LC50", "LC50DM", "IGC50", "LD50", "BCF", "DevTox", "Mutagenicity",
    "BP", "VP", "MP", "Density", "FP", "ST", "TC", "Viscosity", "WS"),
  method = c("consensus", "hc", "sm", "nn", "gc"),
  verbose = TRUE
)
}
\arguments{
\item{query}{character; A SMILES string.}

\item{from}{character; Query identifier. Can be one of: "smiles".}

\item{endpoint}{character; Which predicted endpoint should be returned?
Can be one of "LC50" (Pimephales promelas 96 hr), "LC50DM" (Daphnia magna 48h)
"IGC50" (Tetrahymena pyriformis 48 hr), "LD50" (Oral rat),
"BCF" (Bioaccumulation factor), "DevTox" (Developmental Toxicity),
"Mutagenicity", "BP" (Normal boiling point), "VP" (Vapor pressure at 25°C),
"MP" (Melting point), "Density" (Flash point), "FP" (Density),
"ST" (Surface tension at 25°C), "TC" (Thermal conductivity at 25°C),
"Viscosity" (Viscosity at 25°C) and "WS" (Water solubility at 25°C).
For detail see
\url{https://www.epa.gov/sites/production/files/2018-08/documents/webtest_users_guide.pdf}.}

\item{method}{character; Which prediction method should be used? Can be one
of "consensus" (default), "hc" (hierachical clustering), "sm" (single model),
"nn" (nearest neighbor), "gc" (Group contribution). For detail see
\url{https://www.epa.gov/sites/production/files/2018-08/documents/webtest_users_guide.pdf}.}

\item{verbose}{logical; should a verbose output be printed on the console?}
}
\value{
data.frame
}
\description{
Queries WebTEST, a Web-Services Program to Estimate Toxicity
from Molecular Structure. \url{https://www.epa.gov/sites/production/files/2018-08/documents/webtest_users_guide.pdf}.
}
\examples{
\dontrun{
# might fail if web site is not available
webtest_query("Clc2cc(Cl)ccc2Oc1ccc(Cl)cc1O")
webtest_query("Clc2cc(Cl)ccc2Oc1ccc(Cl)cc1O",
              endpoint = "LC50")
webtest_query("Clc2cc(Cl)ccc2Oc1ccc(Cl)cc1O",
              endpoint = "LC50DM", method = "nn")

# multiple inputs
query = c("CCO", "O=C(O)CNCP(=O)(O)O", "blub")
webtest_query(query)
}

}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
