% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alanwood.R
\name{build_aw_idx}
\alias{build_aw_idx}
\title{Function to build index}
\usage{
build_aw_idx()
}
\value{
a data.frame with three columns: cas, url, and name
}
\description{
Function to build index
}
\seealso{
\code{\link{aw_query}}, \code{\link{aw_idx}}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}
