% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl.R
\name{chembl_query}
\alias{chembl_query}
\title{Query ChEMBL using ChEMBL IDs}
\usage{
chembl_query(query, resource = "molecule")
}
\arguments{
\item{query}{character; a vector of ChEMBL IDs.}

\item{resource}{character; the ChEMBL resource to query.}
}
\value{
The function return a list of lists, where each element of the list
contains a list of respective query results. Results are simplified, if
possible.
}
\description{
Retrieve ChEMBL data using a vector of ChEMBL IDs.
}
\examples{
\donttest {
# might fail if API is not available
chembl_query("CHEMBL1082", resource = "molecule")
chembl_query(c("CHEMBL25", "CHEMBL1082"), resource = "molecule")
chembl_query("CHEMBL771355", resource = "assay")
}
}
