% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl.R
\name{chembl_query}
\alias{chembl_query}
\title{Query ChEMBL using ChEMBL IDs}
\usage{
chembl_query(
  query,
  resource = "molecule",
  cache = FALSE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{query}{character; a vector of ChEMBL IDs.}

\item{resource}{character; the ChEMBL resource to query. Use
\code{chembl_resources()} to see all available resources.}

\item{verbose}{logical; should a verbose output be printed on the console?}
}
\value{
The function return a list of lists, where each element of the list
contains a list of respective query results. Results are simplified, if
possible.
}
\description{
Retrieve ChEMBL data using a vector of ChEMBL IDs.
}
\details{
Each entry in ChEMBL has a unique ID. Data in ChEMBL is organized in
databases called resources. An entry may or may not have a record in a
particular resource. An entry may have a record in more than resources, e.g.
a compound may be present in both the "molecule" and the "drug" resource.
This function queries a vector of ChEMBL IDs from a specific ChEMBL resource.
If you are unsure which ChEMBL resource contains your ChEMBL ID, ese this
function with the \code{"chembl_id_lookup"} resource to find the appropriate
resource for a ChEMBL ID.
}
\note{
Links to the webservice documentation:
\itemize{
     \item \url{https://chembl.gitbook.io/chembl-interface-documentation},
     \item \url{https://www.ebi.ac.uk/chembl/api/data/docs}
}
}
\examples{
\donttest{
# Might fail if API is not available

# Search molecules
chembl_query("CHEMBL1082", resource = "molecule")
chembl_query(c("CHEMBL25", "CHEMBL1082"), resource = "molecule")

# Look up ChEMBL IDs in ChEMBL "resources", returns one resource per query.
chembl_query("CHEMBL771355", "chembl_id_lookup")

# Search assays
chembl_query("CHEMBL771355", resource = "assay")
}
}
\references{
Gaulton, A., Bellis, L. J., Bento, A. P., Chambers, J.,
Davies, M., Hersey, A., ... & Overington, J. P. (2012). ChEMBL: a large-scale
bioactivity database for drug discovery. Nucleic acids research, 40(D1),
D1100-D1107.
}
