% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl.R
\name{chembl_resources}
\alias{chembl_resources}
\title{List ChEMBL Resources}
\usage{
chembl_resources()
}
\description{
Data in ChEMBL is organized in databases called resources. This function
lists available ChEMBL resources.
}
\note{
The list was compiled manually using the following url:
\url{https://chembl.gitbook.io/chembl-interface-documentation/web-services/chembl-data-web-services}
}
\references{
Gaulton, A., Bellis, L. J., Bento, A. P., Chambers, J.,
Davies, M., Hersey, A., ... & Overington, J. P. (2012). ChEMBL: a large-scale
bioactivity database for drug discovery. Nucleic acids research, 40(D1),
D1100-D1107.
}
