% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chooser}
\alias{chooser}
\title{Interactively choose a result from a menu}
\usage{
chooser(x, choices)
}
\arguments{
\item{x}{a character vector}

\item{choices}{If \code{choices = "all"} then the entire vector \code{x} is
used for the menu.  If numeric > 1, only that number of elements from the
start of \code{x} are shown. If \code{choices = 1}, then the first element of
\code{x} is returned without prompting the user.  If \code{NULL} then
\code{x} is returned unchanged.}
}
\value{
a character vector of length 1
}
\description{
In interactive sessions, prompts a user to choose an element of a vector from a menu. Use this for all functions that return multiple possible results such as multiple identifiers or synonyms.
}
\examples{
test <- c("apples", "bananas", "orange", "plum", "peach", "guava", "kumquat")
chooser(test, "all")
chooser(test, 3)
}
