% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{cs_check_key}
\alias{cs_check_key}
\title{Retrieve stored ChemSpider API keys and secure tokens}
\usage{
cs_check_key()
}
\value{
an API key
}
\description{
Look for and retrieve ChemSpider API keys and secure tokens stored in
.Renviron or .Rprofile.
}
\details{
To use the any of the functions in \code{webchem} that access the
  ChemSpider database, you'll need to obtain an API key by registering.
  Register at \url{https://developer.rsc.org/} for an API key. Please respect
  the Terms & Conditions. The Terms & Conditions can be found at
  \url{https://developer.rsc.org/terms}. Rather than store the API key in code, which might be shared, it's best practice to keep it
  hidden. You can store the API key as \code{CHEMSPIDER_KEY=<your key>} in
  .Renviron or as \code{option(chemspider_key = <your key>)} in .Rprofile.
}
\examples{
\dontrun{
cs_check_key()
}
}
\seealso{
\code{\link[usethis]{edit_r_environ}}
  \code{\link[usethis]{edit_r_profile}}
}
