% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{cs_convert}
\alias{cs_convert}
\title{Convert identifiers using Chemspider}
\usage{
cs_convert(query, from, to, apikey)
}
\arguments{
\item{query}{character; query ID.}

\item{from}{character; type of query ID.}

\item{to}{character; type to convert to.}

\item{apikey}{character; your API key.}
}
\value{
Returns a character vector of length one containing the converted
identifier.
}
\description{
Submit an identifier (CSID, SMILES, InChI, InChIKey or Mol) and return an
identifier in another format (CSID, SMILES, InChI, InChIKey or Mol).
}
\details{
Not all conversions are supported. Allowed conversions:
\itemize{
\item CSID <-> InChI
\item CSID <-> InChIKey
\item CSID <-> SMILES
\item CSID -> Mol file
\item InChI <-> InChIKey
\item InChI <-> SMILES
\item InChI <-> Mol file
\item InChIKey <-> Mol file
}
}
\note{
An API key is needed. Register at \url{https://developer.rsc.org/}
for an API key. Please respect the Terms & Conditions. The Terms & Conditions
can be found at \url{https://developer.rsc.org/terms}.
}
\examples{
\dontrun{
apikey <- "<YOUR-API-KEY>"
cs_convert("BQJCRHHNABKAKU-KBQPJGBKSA-N",
  from = "inchikey", to = "csid",
  apikey = apikey
)
cs_convert("BQJCRHHNABKAKU-KBQPJGBKSA-N",
  from = "inchikey", to = "inchi",
  apikey = apikey
)
cs_convert("BQJCRHHNABKAKU-KBQPJGBKSA-N",
  from = "inchikey", to = "mol",
  apikey = apikey
)
cs_convert(160, from = "csid", to = "smiles", apikey = apikey)
}
}
\references{
https://developer.rsc.org/compounds-v1/apis
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}

Tamas Stirling, \email{stirling.tamas@gmail.com}
}
