% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{cs_datasources}
\alias{cs_datasources}
\title{Retrieve ChemSpider data sources}
\usage{
cs_datasources(apikey)
}
\arguments{
\item{apikey}{character; your API key.}
}
\value{
Returns a character vector.
}
\description{
The function returns a vector of available data sources used by ChemSpider.
Many ChemSpider functions allow you to restrict which sources are used to
lookup the requested query. Restricting the sources makes queries faster.
}
\note{
An API key is neeeded. Register at \url{https://developer.rsc.org/}
for an API key. Please respect the Terms & Conditions. The Terms & Conditions
can be found at \url{https://developer.rsc.org/terms}.
}
\examples{
\dontrun{
apikey <- "<YOUR-API-KEY>"
cs_datasources(apikey = apikey)
}
}
\author{
Tamas Stirling, \email{stirling.tamas@gmail.com}
}
