% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{cs_datasources}
\alias{cs_datasources}
\title{Retrieve ChemSpider data sources}
\usage{
cs_datasources(apikey = NULL, verbose = TRUE)
}
\arguments{
\item{apikey}{character; your API key. If NULL (default),
\code{cs_check_key()} will look for it in .Renviron or .Rprofile.}

\item{verbose}{should a verbose output be printed on the console?}
}
\value{
Returns a character vector.
}
\description{
The function returns a vector of available data sources used by ChemSpider.
Some ChemSpider functions allow you to restrict which sources are used to
lookup the requested query. Restricting the sources makes these queries
faster.
}
\note{
An API key is needed. Register at \url{https://developer.rsc.org/}
for an API key. Please respect the Terms & Conditions. The Terms & Conditions
can be found at \url{https://developer.rsc.org/terms}.
}
\examples{
\dontrun{
cs_datasources()
}
}
\references{
\url{https://developer.rsc.org/compounds-v1/apis}
}
