% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{cs_extcompinfo}
\alias{cs_extcompinfo}
\title{Get extended record details by ChemSpider ID}
\usage{
cs_extcompinfo(csid, token, verbose = getOption("verbose"), ...)
}
\arguments{
\item{csid}{character,  ChemSpider ID.}

\item{token}{character; security token.}

\item{verbose}{logical; should a verbose output be printed on the console?}

\item{...}{currently not used.}
}
\value{
a data.frame with entries: 'csid', 'mf' (molecular formula),
'smiles', 'inchi' (non-standard), 'inchikey' (non-standard), 'average_mass',
'mw' (Molecular weight), 'monoiso_mass' (MonoisotopicMass), nominal_mass',
'alogp', 'xlogp', 'common_name' and 'source_url'
}
\description{
Get extended info from ChemSpider, see \url{https://www.chemspider.com/}
}
\note{
A security token is needed. Please register at RSC
\url{https://www.rsc.org/rsc-id/register}
for a security token.

Terms and Conditions: \url{https://developer.rsc.org/terms}

use \code{\link{cs_compinfo}} to retrieve standard inchikey.
}
\examples{
\dontrun{
token <- "<redacted>"
csid <- get_csid("Triclosan")
cs_extcompinfo(csid, token)

csids <- get_csid(c('Aspirin', 'Triclosan'))
cs_compinfo(csids)
}
}
\seealso{
\code{\link{get_csid}} to retrieve ChemSpider IDs,
\code{\link{cs_compinfo}} for extended compound information.
}
