% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{cs_inchi_csid}
\alias{cs_inchi_csid}
\title{Retrieve ChemSpider ID from InChI}
\usage{
cs_inchi_csid(inchi, apikey = NULL)
}
\arguments{
\item{inchi}{character; search term.}

\item{apikey}{character; your API key. If NULL (default),
\code{cs_check_key()} will look for it in .Renviron or .Rprofile.}
}
\value{
Returns a numeric vector of ChemSpider IDs.
}
\description{
Query a single compund by InChI and return a vector of ChemSpider IDs.
}
\note{
An API key is needed. Register at RSC \url{https://developer.rsc.org/}
for an API key. Please respect the Terms & conditions
\url{https://developer.rsc.org/terms}.

This is a low level function and is not exported.
}
\examples{
\dontrun{
cs_inchi_csid(inchi = "InChI=1S/C2H4O2/c1-2(3)4/h1H3,(H,3,4)")
cs_inchi_csid(inchi = c(
"InChI=1S/C2H4O2/c1-2(3)4/h1H3,(H,3,4)",
"InChI=1/C2H4O/c1-2-3/h2H,1H3"))
}
}
\references{
https://developer.rsc.org/compounds-v1/apis
}
\author{
Tamas Stirling, \email{stirling.tamas@gmail.com}
}
