% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{cs_inchikey_csid}
\alias{cs_inchikey_csid}
\title{Retrieve ChemSpider ID from InChIKey}
\usage{
cs_inchikey_csid(inchikey, apikey = NULL)
}
\arguments{
\item{inchikey}{character; search term.}

\item{apikey}{character; your API key. If NULL (default),
\code{cs_check_key()} will look for it in .Renviron or .Rprofile.}
}
\value{
Returns a numeric vector of ChemSpider IDs.
}
\description{
Query a single compund by InChI and return a vector of ChemSpider IDs.
}
\note{
An API key is needed. Register at RSC \url{https://developer.rsc.org/}
for an API key. Please respect the Terms & conditions
\url{https://developer.rsc.org/terms}.

This is a low level function and is not exported.
}
\examples{
\dontrun{
cs_inchikey_csid("QTBSBXVTEAMEQO-UHFFFAOYAR")
cs_inchikey_csid(c("QTBSBXVTEAMEQO-UHFFFAOYAR", "IKHGUXGNUITLKF-UHFFFAOYAB"))
}
}
\references{
https://developer.rsc.org/compounds-v1/apis
}
\author{
Tamas Stirling, \email{stirling.tamas@gmail.com}
}
