% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{cs_inchikey_csid}
\alias{cs_inchikey_csid}
\title{Convert a InChIKey to CSID}
\usage{
cs_inchikey_csid(inchikey, verbose = TRUE, ...)
}
\arguments{
\item{inchikey}{character,  InChIKey}

\item{verbose}{logical; should a verbose output be printed on the console?}

\item{...}{currently not used.}
}
\value{
A CSID.
}
\description{
Convert a InChIKey to CSID
}
\examples{
\donttest{
# might fail if API is not available
cs_inchikey_csid('BQJCRHHNABKAKU-KBQPJGBKSA-N')
}
}
\seealso{
This is a low level function. Please see \code{\link{cs_convert}} for the top level function.
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}
