% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{cs_inchikey_inchi}
\alias{cs_inchikey_inchi}
\title{Convert a InChIKey to InChI}
\usage{
cs_inchikey_inchi(inchikey, token, verbose = TRUE, ...)
}
\arguments{
\item{inchikey}{character,  InChIKey}

\item{token}{character; security token.}

\item{verbose}{logical; should a verbose output be printed on the console?}

\item{...}{currently not used.}
}
\value{
character; InChI
}
\description{
Convert a InChIKey to InChI
}
\note{
A security token is neeeded. Please register at RSC
\url{https://www.rsc.org/rsc-id/register}
for a security token.
}
\examples{
\dontrun{
# Fails because no TOKEN is included
token <- '<YOUR-SECURITY-TOKEN>'
# convert CAS to CSID
cs_inchikey_inchi('BQJCRHHNABKAKU-KBQPJGBKSA-N', token = token)
}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

