% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{cs_smiles_csid}
\alias{cs_smiles_csid}
\title{Retrieve ChemSpider ID from SMILES}
\usage{
cs_smiles_csid(smiles, apikey = NULL)
}
\arguments{
\item{smiles}{character; search term.}

\item{apikey}{character; your API key. If NULL (default), \code{cs_check_key()} will look for it in .Renviron or .Rprofile.}
}
\value{
Returns a list of two elements.
}
\description{
Retrieve ChemSpider ID from SMILES
}
\note{
An API key is needed. Register at RSC
\url{https://developer.rsc.org/}
for an API key.
Please respect the Terms & conditions \url{https://developer.rsc.org/terms}.
}
\examples{
\dontrun{
cs_smiles_csid("CC(O)=O")
cs_smiles_csid(c("CC(O)=O","CC=O"))
}
}
\references{
https://developer.rsc.org/compounds-v1/apis
}
\seealso{
This is a low level function. See \code{\link{cs_convert}}
for the top level function.
}
\author{
Tamas Stirling, \email{stirling.tamas@gmail.com}
}
