% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts.R
\name{cts_from}
\alias{cts_from}
\title{Return a list of all possible ids}
\usage{
cts_from(verbose = TRUE)
}
\arguments{
\item{verbose}{logical; should a verbose output be printed on the console?}
}
\value{
a character vector.
}
\description{
Return a list of all possible ids that can be used in the 'from' argument
}
\details{
See also \url{http://cts.fiehnlab.ucdavis.edu/services}
}
\examples{
\donttest{
cts_from()
}
}
\references{
Wohlgemuth, G., P. K. Haldiya, E. Willighagen, T. Kind, and O. Fiehn 2010The Chemical Translation Service
-- a Web-Based Tool to Improve Standardization of Metabolomic Reports. Bioinformatics 26(20): 2647–2648.
}
\seealso{
\code{\link{cts_convert}}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}
