% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_connect}
\alias{db_connect}
\title{Connect to a database}
\usage{
db_connect(db, version = NULL, ...)
}
\arguments{
\item{db}{character; database name. Must be one of "chembl", "eup", or "foodb".}

\item{version}{character; version of the database. Only applicable for
"chembl". If `NULL` (default), [chembl_check_db_version()] resolves a default
set via .Renviron or .Rprofile.}

\item{...}{Further args passed on to [DBI::dbConnect()]}
}
\value{
an object of class "SQLiteConnection".
}
\description{
This function connects to a database and returns a connection object. 
Currently, it supports the following databases: "chembl", "eup", and "foodb".
}
\examples{
\dontrun{
# Connect to a specific version of the ChEMBL database
con <- db_connect("chembl", version = "37")
# Connect to the EU Pesticides database
con <- db_connect("eup")
# Connect to the FooDB database
con <- db_connect("foodb")
}
}
