% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl.R
\name{db_download_chembl}
\alias{db_download_chembl}
\title{Download ChEMBL database}
\usage{
db_download_chembl(version = "latest", verbose = getOption("verbose"))
}
\arguments{
\item{version}{character, the database release version.}

\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
Downloads the requested database files.
}
\description{
Download a version of the ChEMBL database for offline access.
}
\note{
If a checksum file is available for the requested version it will be
used to check data integrity. To save storage space, webchem only retrieves
those files that are used by the package. If you need other files as well,
please download them manually.
}
\examples{
\dontrun{
db_download_chembl(version = "35", verbose = TRUE)
}
}
\references{
You can find more information about ChEMBL releases at
\url{https://chembl.gitbook.io/chembl-interface-documentation/downloads}
}
