% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etox.R
\name{etox_targets}
\alias{etox_targets}
\title{Get Quality Targets from a ETOX ID}
\usage{
etox_targets(id, verbose = TRUE)
}
\arguments{
\item{id}{character; ETOX ID}

\item{verbose}{logical; print message during processing to console?}
}
\value{
A data.frame with quality targets from the ETOX database.
}
\description{
Query ETOX: Information System Ecotoxicology and Environmental Quality Targets
\url{http://webetox.uba.de/webETOX/index.do} for quality targets
}
\examples{
\dontrun{
id <- get_etoxid('Triclosan')
out <- etox_targets(id)
out[ , c('Substance', 'CAS_NO', 'Country_or_Region', 'Designation',
'Value_Target_LR', 'Unit')]
}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}
\seealso{
\code{\link{get_etoxid}} to retrieve ETOX IDs, \code{\link{etox_basic}} for basic information,
\code{\link{etox_targets}} for quality targets and \code{\link{etox_tests}} for test results
}

