% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eup.R
\name{eup_convert}
\alias{eup_convert}
\title{Convert identifiers in the local EU Pesticides database}
\usage{
eup_convert(
  query,
  from,
  to,
  resource = "active_substances",
  mode = "offline",
  ...
)
}
\arguments{
\item{query}{character; a character of compound identifiers to convert.}

\item{from}{character; the type of identifier to conver from. Can be one of
\code{"substance_id"}, \code{"substance_name"}, or \code{"as_cas_number"}.}

\item{to}{character; the type of identifier to convert to. Can be one of
\code{"substance_id"}, \code{"substance_name"}, or \code{"as_cas_number"}.}

\item{resource}{character; the EU Pesticides resource to query. Can be one of
\code{"active_substances"} or \code{"residues"}.}

\item{mode}{character; the mode of operation. Can be one of \code{"offline"} 
(offline access) or \code{"ws"} (web service access). Currently only offline 
mode is implemented.
... Further args passed on to [DBI::dbConnect()]}
}
\value{
A data frame of converted identifiers, in the same order as
the input \code{query}. If an identifier could not be converted, the
corresponding output will be \code{NA}. If multiple matches are found for a
query, all matches will be returned in separate rows.
}
\description{
Convert identifiers in the local EU Pesticides database
}
\details{
\code{"substance_id"} is the unique identifier for each active
substance in the EU Pesticides database. \code{"substance_name"} is the name
of the active substance, and \code{"as_cas_number"} is the CAS number of the
active substance.
}
\examples{
\dontrun{
# Download database
db_download_eup(verbose = TRUE)

# Query downloaded database
eup_convert(1313, from = "substance_id", to = "substance_name")
eup_convert("Monolinuron", from = "substance_name", to = "as_cas_number")

eup_convert()
}
}
\references{
You can find more information about the EU Pesticides database at
\url{https://food.ec.europa.eu/plants/pesticides/eu-pesticides-database_en}.
}
