% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extractors}
\alias{extractors}
\alias{cas}
\alias{inchikey}
\alias{smiles}
\title{Extract parts from webchem objects}
\usage{
cas(x, ...)

inchikey(x, ...)

smiles(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{currently not used.}
}
\value{
a vector.
}
\description{
Extract parts from webchem objects
}
\references{
Eduard Szöcs, Tamás Stirling, Eric R. Scott, Andreas Scharmüller,
Ralf B. Schäfer (2020). webchem: An R Package to Retrieve Chemical
Information from the Web. Journal of Statistical Software, 93(13).
\doi{10.18637/jss.v093.i13}.
}
