% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extractors}
\alias{extractors}
\alias{cas}
\alias{inchikey}
\alias{smiles}
\title{Extract parts from webchem objects}
\usage{
cas(x, ...)

inchikey(x, ...)

smiles(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{currently not used.}
}
\value{
a vector.
}
\description{
Extract parts from webchem objects
}
