% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{get_cid}
\alias{get_cid}
\title{Retrieve Pubchem Compound ID (CID)}
\usage{
get_cid(
  query,
  from = "name",
  match = c("all", "first", "ask", "na"),
  verbose = TRUE,
  arg = NULL,
  first = NULL,
  ...
)
}
\arguments{
\item{query}{character; search term, one or more compounds.}

\item{from}{character; type of input, can be one of "name" (default), "cid",
"smiles", "inchi", "sdf", "inchikey", "formula", <structure search>, <xref>,
<fast search>. See details for more information.}

\item{match}{character; How should multiple hits be handled?, "all" all
matches are returned, "best" the best matching is returned, "ask" enters an
interactive mode and the user is asked for input, "na" returns NA if multiple
hits are found.}

\item{verbose}{logical; should a verbose output be printed on the console?}

\item{arg}{character; optinal arguments like "name_type=word" to match
individual words.}

\item{first}{deprecated. Use `match` instead.}

\item{...}{currently unused.}
}
\value{
a tibble.
}
\description{
Retrieve compound IDs (CIDs) from PubChem.
}
\details{
<structure search> is assembled as "{\code{substructure} |
\code{superstructure} | \code{similarity} | \code{identity}} / {\code{smiles}
 | \code{inchi} | \code{sdf} | \code{cid}}", e.g.
 \code{from = "substructure/smiles"}.

\code{<xref>} is assembled as "\code{xref}/\{\code{RegistryID} |
\code{RN} | \code{PubMedID} | \code{MMDBID} | \code{ProteinGI},
\code{NucleotideGI} | \code{TaxonomyID} | \code{MIMID} | \code{GeneID} |
\code{ProbeID} | \code{PatentID}\}", e.g. \code{from = "xref/RN"} will query
by CAS RN.

<fast search> is either \code{fastformula} or it is assembled as
"{\code{fastidentity} | \code{fastsimilarity_2d} | \code{fastsimilarity_3d} |
\code{fastsubstructure} | \code{fastsuperstructure}}/{\code{smiles} |
\code{smarts} | \code{inchi} | \code{sdf} | \code{cid}}", e.g.
\code{from = "fastidentity/smiles"}.
}
\note{
Please respect the Terms and Conditions of the National Library of
Medicine, \url{https://www.nlm.nih.gov/databases/download.html} and the data
usage policies of National Center for Biotechnology Information,
\url{https://www.ncbi.nlm.nih.gov/home/about/policies/},
\url{https://pubchemdocs.ncbi.nlm.nih.gov/programmatic-access}.
}
\examples{
\donttest{
# might fail if API is not available
get_cid("Triclosan")
get_cid("Triclosan", arg = "name_type=word")
get_cid("BPGDAMSIGCZZLK-UHFFFAOYSA-N", from = "inchikey")
get_cid("CCCC", from = "smiles")
get_cid("C26H52NO6P", from = "formula")
get_cid("56-40-6", from = "xref/rn")
get_cid(5564, from = "similarity/cid")
get_cid("CCO", from = "similarity/smiles")

# multiple inputs
comp <- c("Triclosan", "Aspirin")
get_cid(comp)

}
}
\references{
Wang, Y., J. Xiao, T. O. Suzek, et al. 2009 PubChem: A Public
Information System for
Analyzing Bioactivities of Small Molecules. Nucleic Acids Research 37:
623–633.

Kim, Sunghwan, Paul A. Thiessen, Evan E. Bolton, et al. 2016
PubChem Substance and Compound Databases. Nucleic Acids Research 44(D1):
D1202–D1213.

Kim, S., Thiessen, P. A., Bolton, E. E., & Bryant, S. H. (2015).
PUG-SOAP and PUG-REST: web services for programmatic access to chemical
information in PubChem. Nucleic acids research, gkv396.
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}
