% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{get_cid}
\alias{get_cid}
\title{Retrieve Pubchem Id (CID)}
\usage{
get_cid(
  query,
  from = c("name", "cid", "sid", "aid", "smiles", "inchi", "inchikey"),
  match = c("all", "first", "ask", "na"),
  verbose = TRUE,
  search_substances = FALSE,
  arg = NULL,
  first = NULL,
  ...
)
}
\arguments{
\item{query}{character; search term.}

\item{from}{character; type of input, can be one of "name" (default), "cid",
"sid", "aid", "smiles", "inchi", "inchikey"}

\item{match}{character; How should multiple hits be handled?, "all" all matches are returned, "best" the best matching is returned, "ask" enters an interactive mode and the user is asked for input, "na" returns NA if multiple hits are found.}

\item{verbose}{logical; should a verbose output be printed on the console?}

\item{search_substances}{logical; If TRUE also searches PubChem SIDs}

\item{arg}{character; optinal arguments like "name_type=word" to match
individual words.}

\item{first}{deprecated. Use `match` instead.}

\item{...}{currently unused.}
}
\value{
a tibble.
}
\description{
Return CompoundID (CID) for a search query using PUG-REST,
see \url{https://pubchem.ncbi.nlm.nih.gov/}.
}
\note{
Please respect the Terms and Conditions of the National Library of
Medicine, \url{https://www.nlm.nih.gov/databases/download.html} and the data
usage policies of National Center for Biotechnology Information,
\url{https://www.ncbi.nlm.nih.gov/home/about/policies/},
\url{https://pubchemdocs.ncbi.nlm.nih.gov/programmatic-access}.
}
\examples{
\donttest{
# might fail if API is not available
get_cid("Triclosan")
get_cid("Triclosan", arg = "name_type=word")
get_cid("BPGDAMSIGCZZLK-UHFFFAOYSA-N", from = "inchikey")
get_cid("CCCC", from = "smiles")

# multiple inputs
comp <- c("Triclosan", "Aspirin")
get_cid(comp)

}
}
\references{
Wang, Y., J. Xiao, T. O. Suzek, et al. 2009 PubChem: A Public
Information System for
Analyzing Bioactivities of Small Molecules. Nucleic Acids Research 37:
623–633.

Kim, Sunghwan, Paul A. Thiessen, Evan E. Bolton, et al. 2016
PubChem Substance and Compound Databases. Nucleic Acids Research 44(D1):
D1202–D1213.

Kim, S., Thiessen, P. A., Bolton, E. E., & Bryant, S. H. (2015).
PUG-SOAP and PUG-REST: web services for programmatic access to chemical
information in PubChem. Nucleic acids research, gkv396.
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}
