% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{get_cid}
\alias{get_cid}
\title{Retrieve Pubchem Id (CID)}
\usage{
get_cid(query, first = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{query}{charachter; search term.}

\item{first}{logical; If TRUE return only first result.}

\item{verbose}{logical; should a verbose output be printed on the console?}

\item{...}{currently not used.}
}
\value{
a character vector.
}
\description{
Return CompoundID (CID) for a search query, see \url{https://pubchem.ncbi.nlm.nih.gov/}.
}
\examples{
\donttest{
# might fail if API is not available
get_cid('Triclosan')

# multiple inputs
comp <- c('Triclosan', 'Aspirin')
sapply(comp, function(x) get_cid(x))
sapply(comp, function(x) get_cid(x, first = TRUE))
}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

