% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{get_csid}
\alias{get_csid}
\title{ChemSpider ID from compound name}
\usage{
get_csid(query, apikey = NULL, control = cs_control())
}
\arguments{
\item{query}{character; search term.}

\item{apikey}{character; your API key. If NULL (default), `cs_check_key()` will look for it in .Renviron or .Rprofile.}

\item{control}{function; see details.}
}
\value{
Returns a list of two elements.
}
\description{
Submit names of compunds and retrieve their ChemSpider IDs (CSIDs).
}
\details{
Control options available for this function are \code{order_by},
\code{order_direction}. See \code{cs_control()} for a full list of valid
values for these control options.
}
\note{
An API key is needed. Register at \url{https://developer.rsc.org/}
for an API key. Please respect the Terms & conditions:
\url{https://developer.rsc.org/terms}.
}
\examples{
\dontrun{
get_csid("triclosan")
get_csid(c("carbamazepine", "naproxene"))
}
}
\references{
https://developer.rsc.org/compounds-v1/apis
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}

Tamas Stirling, \email{stirling.tamas@gmail.com}
}
