% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{get_csid}
\alias{get_csid}
\title{Retrieve ChemSpider ID from a queryId returned by a non-batch POST request}
\usage{
get_csid(postres, headers)
}
\arguments{
\item{postres}{an object of class \code{\link{response()}} containing the
queryId returned by the previous API.}

\item{headers}{list; contains the API key.}
}
\value{
Returns a list of two elements.
}
\description{
Chemspider IDs (CSIDs) are obtained in multiple steps.
First we assemble a POST request.
These requests can take multiple forms and are assembled separately in each
function that returns a CSID.
A dedicated API handles the POST request and returns a queryId.
This queryId is passed to another, more general API, that returns the CSID.
This function communicates with the second API.
}
\note{
An API key is needed. Register at RSC.
\url{https://developer.rsc.org/}
for an API key.
Please respect the Terms & conditions \url{https://developer.rsc.org/terms}.
}
\references{
\url{https://developer.rsc.org/compounds-v1/apis}
}
\author{
Tamas Stirling, \email{stirling.tamas@gmail.com}
}
