% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{get_csid}
\alias{get_csid}
\title{ChemSpider ID from compound name}
\usage{
get_csid(query, apikey, control = cs_control())
}
\arguments{
\item{query}{character; search term.}

\item{apikey}{character; your API key.}

\item{control}{function; see details.}
}
\value{
Returns a list of two elements.
}
\description{
Submit the name of a compund and retrieve the ChemSpider ID (CSID) of the
compound.
}
\details{
Control options available for this function are \code{orderBy},
\code{orderDirection}. See \code{cs_control()} for a full list of valid values
for these control options.
}
\note{
An API key is needed. Register at RSC.
\url{https://developer.rsc.org/}
for an API key.
Please respect the Terms & conditions \url{https://developer.rsc.org/terms}.
}
\examples{
\dontrun{
apikey <- "<YOUR-API-KEY>"
get_csid("Triclosan", apikey = apikey)
}
}
\references{
https://developer.rsc.org/compounds-v1/apis
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}

Tamas Stirling, \email{stirling.tamas@gmail.com}
}
