% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemspider.R
\name{get_csid}
\alias{get_csid}
\title{Retrieve ChemSpider ID}
\usage{
get_csid(query, apikey, orderBy = "recordId",
  orderDirection = "ascending", ...)
}
\arguments{
\item{query}{character; search term.}

\item{apikey}{character; your API key.}

\item{orderBy}{character; used to to specify the sort order for the results.
Valid values for orderBy are recordId, massDefect, molecularWeight,
referenceCount, dataSourceCount, pubMedCount, rscCount.}

\item{orderDirection}{character; used to to specify the sort order for the results.
Valid values for orderDirection are ascending, descending.}

\item{...}{currently not used.}
}
\value{
Returns a list of two elements.
}
\description{
Return Chemspider ID (CSID) for a search query, see \url{https://www.chemspider.com/}.
}
\note{
An API key is neeeded. Register at RSC.
\url{https://developer.rsc.org/}
for an API key.
Please respect the Terms & conditions \url{https://developer.rsc.org/terms}.
}
\examples{
\dontrun{
apikey <- '<YOUR-API-KEY>'
get_csid("Triclosan", apikey = apikey)
}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}

Tamas Stirling, \email{stirling.tamas@gmail.com}
}
