% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webchem.R
\docType{data}
\name{jagst}
\alias{jagst}
\title{Organic plant protection products in the river Jagst / Germany in 2013}
\format{A data frame with 442 rows and 4 variables:
\describe{
  \item{date}{sampling data}
  \item{substance}{substance names}
  \item{value}{concentration in ug/L}
  \item{qual}{qualifier, indicating values < LOQ}
}}
\source{
\url{http://jdkfg.lubw.baden-wuerttemberg.de/servlet/is/300/}
}
\usage{
jagst
}
\description{
This dataset comprises environmental monitoring data of organic plant protection products
in the year 2013 in the river Jagst, Germany.
The data is publicly available and can be retrieved from the
LUBW Landesanstalt für Umwelt, Messungen und Naturschutz Baden-Württemberg.
It has been preprocessed and comprises measurements of 34 substances.
Substances without detects have been removed.
on 13 sampling occasions.
Values are given in ug/L.
}
\keyword{datasets}
