% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webchem-package.R
\docType{data}
\name{lc50}
\alias{lc50}
\title{Acute toxicity data from U.S. EPA ECOTOX}
\format{A data frame with 124 rows and 2 variables:
\describe{
  \item{cas}{CAS registry number}
  \item{value}{LC50value}
}}
\source{
\url{http://cfpub.epa.gov/ecotox/}
}
\usage{
lc50
}
\description{
This dataset comprises acute ecotoxicity data of 124 insecticides.
The data is publicly available and can be retrieved from the EPA ECOTOX database
(\url{http://cfpub.epa.gov/ecotox/})
It comprises acute toxicity data (D. magna, 48h, Laboratory, 48h) and has been
preprocessed (remove non-insecticides, aggregate multiple value, keep only numeric data etc).
}
\keyword{datasets}
