% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nist.R
\name{nist_ri}
\alias{nist_ri}
\title{Retrieve retention indices from NIST}
\usage{
nist_ri(cas, type = c("kovats", "linear", "alkane", "lee"),
  polarity = c("polar", "non-polar"), temp_prog = c("isothermal",
  "ramp", "custom"))
}
\arguments{
\item{cas}{CAS numbers either as numeric or formatted correctly with hyphens.}

\item{type}{Retention index type. One of `"kovats"`, `"linear"`, `"alkane"`, or `"lee"`. "kovats" returns Kovats retention indices, "linear" returns Van den Dool and Kratz retention indices, "alkane" returns normal alkane retention indices, and "lee" returns Lee retention indices. Details about how these are calculated here: [https://webbook.nist.gov/chemistry/gc-ri/].}

\item{polarity}{Column polarity. One of "polar" or "non-polar" to get RIs calculated for polar or non-polar columns.}

\item{temp_prog}{Temperature program. One of "isothermal", "ramp", or "custom".}
}
\value{
a table of literature RIs
}
\description{
This function scrapes NIST for literature retention indices given CAS numbers as an input.
}
\examples{
myRIs <- nist_ri(c("78-70-6", "13474-59-4"), "linear", "non-polar", "ramp")

}
\seealso{
\code{\link{is.cas}} \code{\link{as.cas}}
}
