% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{pc_extract}
\alias{pc_extract}
\title{Extract data from PubChem content pages}
\usage{
pc_extract(page, section)
}
\arguments{
\item{page}{list; a list of PubChem content pages.}

\item{section}{character; the lowest level section of the data to be
accessed.}
}
\value{
A tibble of chemical information with references.
}
\description{
This function takes a list of PubChem content pages, and extracts the
required information from them.
}
\details{
When you look at a PubChem content page, you can see that chemical
information is organised into sections, subsections, etc. The chemical data
live at the lowest levels of these sections. Use this function to extract the
lowest level information from PubChem content pages, e.g. IUPAC Name, Boiling
Point, Lower Explosive Limit (LEL).

The \code{section} argument is not case sensitive, but it is
sensitive to typing errors, and requires the full name of the section as it
is printed on the content page. The PubChem Table of Contents Tree can also
be found at \url{https://pubchem.ncbi.nlm.nih.gov/classification/#hid=72}.
}
\note{
Please respect the Terms and Conditions and data usage policies of the
service provider and the data providers:
\itemize{
\item National Library of Medicine: \cr
\url{https://www.nlm.nih.gov/databases/download.html}
\item National Center for Biotechnology Information: \cr
\url{https://www.ncbi.nlm.nih.gov/home/about/policies/}
\url{https://pubchemdocs.ncbi.nlm.nih.gov/programmatic-access}
\item data providers: \cr
\url{https://pubchem.ncbi.nlm.nih.gov/sources/}}
}
\examples{
\donttest{
# Import all experimental properties for acetic acid and citric acid
pages <- pc_page(c(176, 311), "experimental properties")
# Extract pKa data
pc_extract(pages, "pka")

# Import all toxicity data for benzene
pages <- pc_page(241, "toxicity")
# Extract exposure routes
pc_extract(pages, "exposure routes")$String

# Import all hazards identification data for formaldehyde
pages <- pc_page(712, "hazards identification")
# Extract
pc_extract(pages, "EPA Hazardous Waste Number")$String
}
}
\author{
Tamas Stirling, \email{stirling.tamas@gmail.com}
}
