% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{ping}
\alias{pubchem_ping}
\title{Ping an API used in webchem to see if it's working.}
\usage{
pubchem_ping(what = "status", ...)
}
\arguments{
\item{what}{(character) One of 'status' (default), 'content', or an HTTP status
code. If status, we just check that the HTTP status code is 200, or similar
signifying the service is up. If content, we do a simple, quick check to
determine if returned content matches what's expected. If an HTTP status
code, it must match an appropriate code. See \code{\link{status_codes}}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}
}
\value{
A logical, TRUE or FALSE
}
\description{
Ping an API used in webchem to see if it's working.
}

