% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{cs_web_ping}
\alias{ping}
\alias{pubchem_ping}
\title{Ping an API used in webchem to see if it's working.}
\usage{
pubchem_ping(...)

cs_web_ping(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}
}
\value{
A logical, TRUE or FALSE
}
\description{
Ping an API used in webchem to see if it's working.
}

