% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppdb.R
\docType{data}
\name{ppdb_idx}
\alias{ppdb_idx}
\title{PPDB search index}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{cas}{cas}
  \item{link}{matched link}
}}
\source{
\url{http://sitem.herts.ac.uk/aeru/iupac/search.htm}
}
\usage{
ppdb_idx
}
\description{
A dataset containing the matched strings and links of the PPDB  \url{http://sitem.herts.ac.uk/aeru/iupac/search.htm}.
This dataset has been created using code{\link{ppdb_buildidx}}
}
\details{
Retrieved using \code{\link{ppdb_idx}} on 11th October 2015.
}
\seealso{
\code{\link{ppdb_buildidx}}
}
\keyword{datasets}

