% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppdb.R
\name{ppdb_parse}
\alias{ppdb_parse}
\title{Parse a HTML source from PPDB.}
\usage{
ppdb_parse(source, verbose = TRUE)
}
\arguments{
\item{source}{an object of class \code{xml_document} as returned by  \code{\link[xml2]{read_html}}.}

\item{verbose}{logical; print message during processing to console?}
}
\value{
A list of 11 data.frames : ec_regulation, approved_in, general, parents, fate,
deg, soil, metab, etox, names and source_url.
}
\description{
This function parses a (substance) html from the website into an R object.
Earlier versions allowed also to search and download the database.
However, this is explicitly against the terms and conditions of use [link removed on request].
On request we also removed all links to the website / database.
}
\note{
Please read the Terms and Conditions for use [link removed on request] and
the Copyright statement [link removed on request].

This function only parses a html. Saving (or downloading) substantial parts
from the database is explicitly against the terms and conditions and copyright of use
[link removed on request].
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}
\references{
[Reference removed on request.]
}

