% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodscents.R
\name{tgsc_percept}
\alias{tgsc_percept}
\title{Scrape \url{thegoodsscentscompany.com} for aromas of compounds}
\usage{
tgsc_percept(CAS, odor = TRUE, flavor = FALSE, ...)
}
\arguments{
\item{CAS}{character; CAS number to search by. See \code{\link{is.cas}} for
correct formatting.}

\item{odor}{logical; TRUE by default, if TRUE, returns odor data from web}

\item{flavor}{logical; FALSE by default, if TRUE, returns flavor data from web}

\item{...}{Currently not used}
}
\value{
Returns a list of character vectors the found odors and/or flavors
or an NA if not found.
}
\description{
Scrape \url{thegoodsscentscompany.com} for aromas of compounds
}
\examples{
\donttest{
tgsc_percept('66-25-1')
}
}
\author{
Josh Morimoto, \email{joshua.morimoto@tufts.edu}
}
