% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nist.R
\name{tidy_ritable}
\alias{tidy_ritable}
\title{Tidier for webscraped RI ri_xml}
\usage{
tidy_ritable(ri_xml)
}
\arguments{
\item{ri_xml}{captured by `get_ri_xml`}
}
\value{
a single table
}
\description{
Tidier for webscraped RI ri_xml
}
