% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webchem-defunct.R
\docType{package}
\name{webchem-defunct}
\alias{allanwood}
\alias{webchem-defunct}
\alias{webchem-defunct-package}
\title{Defunct function(s) in the webchem package}
\usage{
allanwood(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are defunct and no longer available.
}
\details{
Defunct functions are:
\tabular{rl}{
  \code{allanwood} \cr
}
}

