% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webchem-defunct.R
\name{webchem-defunct}
\alias{webchem-defunct}
\alias{ppdb_query}
\alias{ppdb_parse}
\alias{ppdb}
\alias{cir}
\alias{pp_query}
\alias{cs_prop}
\alias{ci_query}
\alias{pan_query}
\title{Defunct function(s) in the webchem package}
\usage{
ppdb_query()

ppdb_parse()

ppdb()

cir()

pp_query()

cs_prop()

ci_query()

pan_query()
}
\description{
These functions are defunct and no longer available.
}
