% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webchem-defunct.R
\name{webchem-defunct}
\alias{csid_compinfo}
\alias{csid_extcompinfo}
\alias{webchem-defunct}
\title{Defunct function(s) in the webchem package}
\usage{
csid_extcompinfo()

csid_compinfo()
}
\description{
These functions are defunct and no longer available.
}
\details{
Defunct functions are:
\tabular{rl}{
  \code{csid_extcompinfo} \cr
  \code{csid_extcompinfo} \cr
}
}

