% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webchem-defunct.R
\name{webchem-defunct}
\alias{allanwood}
\alias{webchem-defunct}
\title{Defunct function(s) in the webchem package}
\usage{
allanwood()
}
\description{
These functions are defunct and no longer available.
}
\details{
Defunct functions are:
\tabular{rl}{
  \code{allanwood} \cr
}
}

