% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webchem-defunct.R
\name{webchem-defunct}
\alias{cir}
\alias{ppdb}
\alias{ppdb_query}
\alias{webchem-defunct}
\title{Defunct function(s) in the webchem package}
\usage{
ppdb_query()

ppdb()

cir()
}
\description{
These functions are defunct and no longer available.
}
\details{
Defunct functions are:
\tabular{rl}{
  \code{ppdb_query} \cr
  \code{ppdb} \cr
  \code{cir} \cr
}
}

