% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webchem-deprecated.R
\name{webchem-deprecated}
\alias{cir_query}
\alias{csid_compinfo}
\alias{csid_extcompinfo}
\alias{ppdb_query}
\alias{webchem-deprecated}
\title{Deprecated function(s) in the webchem package}
\usage{
ppdb_query(...)

cir_query(...)

csid_extcompinfo(...)

csid_compinfo(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the webchem package.  They may eventually be completely
removed.
}
\details{
Deprecated functions are:
\tabular{rl}{
  \code{ppdb_query} \tab is now a synonym for \code{\link{ppdb}}\cr
  \code{cir_query} \tab is now a synonym for \code{\link{cir}}\cr
  \code{csid_extcompinfo} \tab is now a synonym for \code{\link{cs_extcompinfo}}\cr
  \code{csid_compinfo} \tab is now a synonym for \code{\link{cs_compinfo}}\cr
}
}

