% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/webchem-deprecated.R
\docType{package}
\name{webchem-deprecated}
\alias{allanwood}
\alias{webchem-deprecated}
\alias{webchem-deprecated-package}
\title{Deprecated function(s) in the webchem package}
\usage{
allanwood(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the webchem package.  They may eventually be completely
removed.
}
\section{Details}{

\tabular{rl}{
  \code{allanwood} \tab now a synonym for \code{\link{alanwood}}\cr
}
}

