% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webchem-deprecated.R
\name{webchem-deprecated}
\alias{webchem-deprecated}
\alias{cid_compinfo}
\alias{aw_query}
\title{Deprecated function(s) in the webchem package}
\usage{
cid_compinfo(...)

aw_query(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the webchem package.  They may eventually be completely
removed.
}
\details{
Deprecated functions are:
\tabular{rl}{
  \code{pc_prop} \tab was formerly \code{\link{cid_compinfo}}\cr
  \code{bcpc_query} \tab was formerly \code{\link{aw_query}}\cr
}
}
