% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webchem-deprecated.R
\name{webchem-deprecated}
\alias{alanwood}
\alias{chemid}
\alias{cid_compinfo}
\alias{cir}
\alias{get_cid}
\alias{pan}
\alias{physprop}
\alias{ppdb}
\alias{webchem-deprecated}
\title{Deprecated function(s) in the webchem package}
\usage{
ppdb(...)

cir(...)

pan(...)

alanwood(...)

physprop(...)

get_cid(...)

cid_compinfo(...)

chemid(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the webchem package.  They may eventually be completely
removed.
}
\details{
Deprecated functions are:
\tabular{rl}{
  \code{ppdb_query} \tab is now a synonym for \code{\link{ppdb}}\cr
  \code{cir_query} \tab is now a synonym for \code{\link{cir}}\cr
  \code{pan_query} \tab is now a synonym for \code{\link{pan}}\cr
  \code{aw_query} \tab is now a synonym for \code{\link{alanwood}}\cr
  \code{pp_query} \tab is now a synonym for \code{\link{physprop}}\cr
  \code{get_pcid} \tab is now a synonym for \code{\link{get_cid}}\cr
  \code{pc_compinfo} \tab is now a synonym for \code{\link{cid_compinfo}}\cr
}
}

