% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webchem-package.R
\docType{package}
\name{webchem}
\alias{webchem}
\title{webchem: An R package to retrieve chemical information from the web.}
\description{
webchem: An R package to retrieve chemical information from the web.
}
